Plutarch

Documentation

data ((a :: PType) :--> (b :: PType)) (s :: S) infixr 0

type ClosedTerm (a :: PType) = forall (s :: S). Term s a

Closed* terms with no free variables.

compile :: Config -> ClosedTerm a -> Either Text Script

Compile a (closed) Plutus Term to a usable script

type Dig = Digest Blake2b_160

hashTerm :: Config -> ClosedTerm a -> Either Text Dig

papp :: HasCallStack => Term s (a :--> b) -> Term s a -> Term s b

Lambda Application.

pdelay :: Term s a -> Term s (PDelayed a)

Plutus 'delay', used for laziness.

data PDelayed (a :: PType) (s :: S)

perror :: Term s a

Plutus 'error'.

When using this explicitly, it should be ensured that the containing term is delayed, avoiding premature evaluation.

pforce :: Term s (PDelayed a) -> Term s a

Plutus 'force', used to force evaluation of PDelayed terms.

phoistAcyclic :: HasCallStack => ClosedTerm a -> Term s a

plet :: Term s a -> (Term s a -> Term s b) -> Term s b

Let bindings.

This is appoximately a shorthand for a lambda and application:

plet v f == papp (plam f) v

But sufficiently small terms in WHNF may be inlined for efficiency.

data Term (s :: S) (a :: PType)

data S

type PType = S -> Type

Shorthand for Plutarch types.

class PlutusType (a :: PType)

type family PInner a :: PType

pcon :: PlutusType a => a s -> Term s a

Construct a Plutarch Term via a Haskell datatype

pmatch :: PlutusType a => Term s a -> (a s -> Term s b) -> Term s b

Pattern match over Plutarch Terms via a Haskell datatype

type PCon = PlutusType

Deprecated: Use PlutusType

type PMatch = PlutusType

Deprecated: Use PlutusType

printTerm :: HasCallStack => Config -> ClosedTerm a -> String

Prettyprint a Term via the PLC pretty printer

TODO: Heavily improve. It's unreadable right now.

We could convert the de Bruijn indices into names with:

show . prettyPlcReadableDef . (\(Right p) -> p) . Scripts.mkTermToEvaluate . compile $ term

printScript :: Script -> String

Prettyprint a compiled Script via the PLC pretty printer

(#$) :: HasCallStack => Term s (a :--> b) -> Term s a -> Term s b infixr 0

Low precedence infixr synonym of papp, to be used like $, in combination with #. e.g.:

>>> f # x #$ g # y # z
f x (g y z)

(#) :: HasCallStack => Term s (a :--> b) -> Term s a -> Term s b infixl 8

High precedence infixl synonym of papp, to be used like function juxtaposition. e.g.:

>>> f # x # y
f x y

pinl :: Term s a -> (Term s a -> Term s b) -> Term s b

pto :: Term s a -> Term s (PInner a)

Safely coerce from a Term to it's PInner representation.

pfix :: Term s (((a :--> b) :--> (a :--> b)) :--> (a :--> b))

Fixpoint recursion. Used to encode recursive functions.

Example:

iterateN' ::
 Term s (PInteger :--> (a :--> a) :--> a :--> a) ->
 Term s PInteger ->
 Term s (a :--> a) ->
 Term s a
iterateN' self n f x =
  pif (n #== 0) x (self # n - 1 #$ f x)

iterateN :: Term s (PInteger :--> (a :--> a) :--> a :--> a)
iterateN = pfix #$ plam iterateN'

Further examples can be found in examples/Recursion.hs

data POpaque s

An Arbitrary Term with an unknown type

Constructors

POpaque (Term s POpaque)

popaque :: Term s a -> Term s POpaque

Erase the type of a Term

plam :: forall c. PLamN a b s => (Term s c -> a) -> Term s (c :--> b)

newtype TermCont :: forall (r :: PType). S -> Type -> Type where

Constructors

TermCont :: forall r s a. ((a -> Term s r) -> Term s r) -> TermCont @r s a

hashOpenTerm :: Term s a -> TermCont s Dig

runTermCont :: TermCont @r s a -> (a -> Term s r) -> Term s r

unTermCont :: TermCont @a s (Term s a) -> Term s a

data Config

Constructors

Config

Fields

tracingMode :: TracingMode

data TracingMode

Constructors

NoTracing

DoTracing

DetTracing

pgetConfig :: (Config -> Term s a) -> Term s a

defaultConfig :: Config

newtype PForall (b :: a -> PType) s

Constructors

PForall (forall (x :: a). Term s (b x))

data PSome (b :: a -> PType) s

Constructors

forall (x :: a). PSome (Term s (b x))

newtype PScottEncoded a r s

Constructors

PScottEncoded (Term s (ScottFn (ScottList a r) r))

data PlutusTypeScott

data PlutusTypeNewtype

class (PInner a ~ DerivedPInner (DPTStrat a) a, PlutusTypeStrat (DPTStrat a), PlutusTypeStratConstraint (DPTStrat a) a, PlutusType a) => DerivePlutusType (a :: PType)

type family DPTStrat a :: Type

class (forall t. PCovariant'' t => PCovariant'' (a t)) => PCovariant a

type family PCovariant' a :: Constraint

class (forall t. PCovariant'' t => PContravariant'' (a t)) => PContravariant a

type family PContravariant' a :: Constraint

class (forall t. PVariant'' t => PVariant'' (a t)) => PVariant a

type family PVariant' a :: Constraint