Plutarch.Api.V1

Contexts

newtype PScriptContext (s :: S)

Script context consists of the script purpose and the pending transaction info.

Constructors

PScriptContext (Term s (PDataRecord '["txInfo" := PTxInfo, "purpose" := PScriptPurpose]))

newtype PTxInfo (s :: S)

A pending transaction. This is the view as seen by the validator script.

Constructors

PTxInfo (Term s (PDataRecord '["inputs" := PBuiltinList (PAsData PTxInInfo), "outputs" := PBuiltinList (PAsData PTxOut), "fee" := PValue 'Sorted 'Positive, "mint" := PValue 'Sorted 'NoGuarantees, "dcert" := PBuiltinList (PAsData PDCert), "wdrl" := PBuiltinList (PAsData (PTuple PStakingCredential PInteger)), "validRange" := PPOSIXTimeRange, "signatories" := PBuiltinList (PAsData PPubKeyHash), "datums" := PBuiltinList (PAsData (PTuple PDatumHash PDatum)), "id" := PTxId]))

data PScriptPurpose (s :: S)

The purpose of the script that is currently running

Constructors

PMinting (Term s (PDataRecord '["_0" := PCurrencySymbol]))

PSpending (Term s (PDataRecord '["_0" := PTxOutRef]))

PRewarding (Term s (PDataRecord '["_0" := PStakingCredential]))

PCertifying (Term s (PDataRecord '["_0" := PDCert]))

Script

newtype PDatum (s :: S)

Constructors

PDatum (Term s PData)

newtype PDatumHash (s :: S)

Constructors

PDatumHash (Term s PByteString)

newtype PRedeemer (s :: S)

Constructors

PRedeemer (Term s PData)

newtype PRedeemerHash (s :: S)

Constructors

PRedeemerHash (Term s PByteString)

newtype PStakeValidatorHash (s :: S)

Constructors

PStakeValidatorHash (Term s PByteString)

newtype PValidatorHash (s :: S)

Constructors

PValidatorHash (Term s PByteString)

Script Utils

validatorHash :: Validator -> ValidatorHash

Hash a Validator, with the correct prefix for Plutus V1

mintingPolicySymbol :: MintingPolicy -> CurrencySymbol

Hash a MintingPolicy, with the correct prefix for Plutus V1

stakeValidatorHash :: StakeValidator -> StakeValidatorHash

Hash a StakeValidator, with the correct prefix for Plutus V1

scriptHash :: Script -> ScriptHash

Hash a Script, with the correct prefix for Plutus V1

datumHash :: Datum -> DatumHash

Hash a Datum.

redeemerHash :: Redeemer -> RedeemerHash

Hash a Redeemer.

dataHash :: ToData a => a -> BuiltinByteString

Hash the data encoded representation of given argument.

mkValidator :: HasCallStack => Config -> ClosedTerm PValidator -> Validator

Compile a Validator

mkStakeValidator :: HasCallStack => Config -> ClosedTerm PStakeValidator -> StakeValidator

Compile a StakeValidator

mkMintingPolicy :: HasCallStack => Config -> ClosedTerm PMintingPolicy -> MintingPolicy

Compile a MintingPolicy

type PValidator = PData :--> (PData :--> (PScriptContext :--> POpaque))

a Validator Term

type PMintingPolicy = PData :--> (PScriptContext :--> POpaque)

a MintingPolicy Term

type PStakeValidator = PData :--> (PScriptContext :--> POpaque)

a StakeValidator Term

Value

newtype PValue (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S)

Constructors

PValue (Term s (PMap keys PCurrencySymbol (PMap keys PTokenName PInteger)))

newtype PCurrencySymbol (s :: S)

Constructors

PCurrencySymbol (Term s PByteString)

newtype PTokenName (s :: S)

Constructors

PTokenName (Term s PByteString)

data KeyGuarantees

Constructors

Sorted

Unsorted

data AmountGuarantees

Constructors

NoGuarantees

NonZero

Positive

Crypto

newtype PPubKeyHash (s :: S)

Constructors

PPubKeyHash (Term s PByteString)

pubKeyHash :: PubKey -> PubKeyHash

DCert

data PDCert (s :: S)

Constructors

PDCertDelegRegKey (Term s (PDataRecord '["_0" := PStakingCredential]))

PDCertDelegDeRegKey (Term s (PDataRecord '["_0" := PStakingCredential]))

PDCertDelegDelegate (Term s (PDataRecord '["_0" := PStakingCredential, "_1" := PPubKeyHash]))

PDCertPoolRegister (Term s (PDataRecord '["_0" := PPubKeyHash, "_1" := PPubKeyHash]))

PDCertPoolRetire (Term s (PDataRecord '["_0" := PPubKeyHash, "_1" := PInteger]))

PDCertGenesis (Term s (PDataRecord '[]))

PDCertMir (Term s (PDataRecord '[]))

Time

data PPOSIXTime (s :: S)

type PPOSIXTimeRange = PInterval PPOSIXTime

Interval

newtype PInterval a (s :: S)

Constructors

PInterval (Term s (PDataRecord '["from" := PLowerBound a, "to" := PUpperBound a]))

newtype PLowerBound a (s :: S)

Constructors

PLowerBound (Term s (PDataRecord '["_0" := PExtended a, "_1" := PClosure]))

newtype PUpperBound a (s :: S)

Constructors

PUpperBound (Term s (PDataRecord '["_0" := PExtended a, "_1" := PClosure]))

data PExtended a (s :: S)

Constructors

PNegInf (Term s (PDataRecord '[]))

PFinite (Term s (PDataRecord '["_0" := a]))

PPosInf (Term s (PDataRecord '[]))

type PClosure = PBool

Address

data PCredential (s :: S)

Constructors

PPubKeyCredential (Term s (PDataRecord '["_0" := PPubKeyHash]))

PScriptCredential (Term s (PDataRecord '["_0" := PValidatorHash]))

data PStakingCredential (s :: S)

Constructors

PStakingHash (Term s (PDataRecord '["_0" := PCredential]))

PStakingPtr (Term s (PDataRecord '["_0" := PInteger, "_1" := PInteger, "_2" := PInteger]))

newtype PAddress (s :: S)

Constructors

PAddress (Term s (PDataRecord '["credential" := PCredential, "stakingCredential" := PMaybeData PStakingCredential]))

Tx

newtype PTxOutRef (s :: S)

Reference to a transaction output with a index referencing which of the outputs is being referred to.

Constructors

PTxOutRef (Term s (PDataRecord '["id" := PTxId, "idx" := PInteger]))

newtype PTxOut (s :: S)

A transaction output. This consists of a target address, value and maybe a datum hash

Constructors

PTxOut (Term s (PDataRecord '["address" := PAddress, "value" := PValue 'Sorted 'Positive, "datumHash" := PMaybeData PDatumHash]))

newtype PTxId (s :: S)

Constructors

PTxId (Term s (PDataRecord '["_0" := PByteString]))

newtype PTxInInfo (s :: S)

A input of the pending transaction.

Constructors

PTxInInfo (Term s (PDataRecord '["outRef" := PTxOutRef, "resolved" := PTxOut]))

AssocMap

newtype PMap (keysort :: KeyGuarantees) (k :: PType) (v :: PType) (s :: S)

Constructors

PMap (Term s (PBuiltinMap k v))

Others

data PMaybeData a (s :: S)

Data encoded Maybe type. Used in various ledger api types.

Constructors

PDJust (Term s (PDataRecord '["_0" := a]))

PDNothing (Term s (PDataRecord '[]))

type PTuple a b = PDataSum '['["_0" := a, "_1" := b]]

Utility functions

ptuple :: Term s (PAsData a :--> (PAsData b :--> PTuple a b))