Plutarch.Api.V1.AssocMap

Documentation

newtype PMap (keysort :: KeyGuarantees) (k :: PType) (v :: PType) (s :: S)

Constructors

PMap (Term s (PBuiltinMap k v))

data KeyGuarantees

Constructors

Sorted

Unsorted

Creation

pempty :: Term s (PMap 'Sorted k v)

Construct an empty PMap.

psingleton :: (PIsData k, PIsData v) => Term s (k :--> (v :--> PMap 'Sorted k v))

Construct a singleton PMap with the given key and value.

psingletonData :: Term s (PAsData k :--> (PAsData v :--> PMap 'Sorted k v))

Construct a singleton PMap with the given data-encoded key and value.

pinsert :: (POrd k, PIsData k, PIsData v) => Term s (k :--> (v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Insert a new key/value pair into the map, overiding the previous if any.

pinsertData :: (POrd k, PIsData k) => Term s (PAsData k :--> (PAsData v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Insert a new data-encoded key/value pair into the map, overiding the previous if any.

pdelete :: (POrd k, PIsData k) => Term s (k :--> (PMap 'Sorted k v :--> PMap 'Sorted k v))

Delete a key from the map.

pfromAscList :: (POrd k, PIsData k, PIsData v) => Term s (PBuiltinMap k v :--> PMap 'Sorted k v)

Construct a PMap from a list of key-value pairs, sorted by ascending key data.

passertSorted :: (POrd k, PIsData k, PIsData v) => Term s (PMap _ k v :--> PMap 'Sorted k v)

Assert the map is properly sorted.

pforgetSorted :: Term s (PMap 'Sorted k v) -> Term s (PMap g k v)

Forget the knowledge that keys were sorted.

Lookups

plookup :: (PIsData k, PIsData v) => Term s (k :--> (PMap _ k v :--> PMaybe v))

Look up the given key in a PMap.

plookupData :: (PIsData k, PIsData v) => Term s (PAsData k :--> (PMap _ k v :--> PMaybe (PAsData v)))

Look up the given key data in a PMap.

pfindWithDefault :: (PIsData k, PIsData v) => Term s (v :--> (k :--> (PMap _ k v :--> v)))

Look up the given key in a PMap, returning the default value if the key is absent.

pfoldAt :: (PIsData k, PIsData v) => Term s (k :--> (r :--> ((PAsData v :--> r) :--> (PMap _ k v :--> r))))

Look up the given key in a PMap; return the default if the key is absent or apply the argument function to the value data if present.

pnull :: Term s (PMap _ k v :--> PBool)

Tests whether the map is empty.

Folds

pall :: PIsData v => Term s ((v :--> PBool) :--> (PMap _ k v :--> PBool))

Tests if all values in the map satisfy the given predicate.

pany :: PIsData v => Term s ((v :--> PBool) :--> (PMap _ k v :--> PBool))

Tests if anu value in the map satisfies the given predicate.

Filters and traversals

pfilter :: PIsData v => Term s ((v :--> PBool) :--> (PMap g k v :--> PMap g k v))

Filters the map so it contains only the values that satisfy the given predicate.

pmap :: (PIsData a, PIsData b) => Term s ((a :--> b) :--> (PMap g k a :--> PMap g k b))

Applies a function to every value in the map, much like map.

pmapData :: Term s ((PAsData a :--> PAsData b) :--> (PMap g k a :--> PMap g k b))

pmapMaybe :: (PIsData a, PIsData b) => Term s ((a :--> PMaybe b) :--> (PMap g k a :--> PMap g k b))

Maps and filters the map, much like mapMaybe.

pmapMaybeData :: Term s ((PAsData a :--> PMaybe (PAsData b)) :--> (PMap g k a :--> PMap g k b))

Combining

pdifference :: (PIsData k, PIsData a, PIsData b) => Term s (PMap g k a :--> (PMap _ k b :--> PMap g k a))

Difference of two maps. Return elements of the first map not existing in the second map.

punionWith :: (POrd k, PIsData k, PIsData v) => Term s ((v :--> (v :--> v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Combine two PMaps applying the given function to any two values that share the same key.

punionWithData :: (POrd k, PIsData k) => Term s ((PAsData v :--> (PAsData v :--> PAsData v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Combine two PMaps applying the given function to any two data-encoded values that share the same key.

Orphan instances

(PConstantData k, PConstantData v, Ord k) => PConstantDecl (Map k v)

Instance details

Associated Types

type PConstantRepr (Map k v)

type PConstanted (Map k v) :: PType

Methods

pconstantToRepr :: Map k v -> PConstantRepr (Map k v)

pconstantFromRepr :: PConstantRepr (Map k v) -> Maybe (Map k v)