Plutarch.Api.V1.Tx

Documentation

newtype PTxOutRef (s :: S)

Reference to a transaction output with a index referencing which of the outputs is being referred to.

Constructors

PTxOutRef (Term s (PDataRecord '["id" := PTxId, "idx" := PInteger]))

newtype PTxOut (s :: S)

A transaction output. This consists of a target address, value and maybe a datum hash

Constructors

PTxOut (Term s (PDataRecord '["address" := PAddress, "value" := PValue 'Sorted 'Positive, "datumHash" := PMaybeData PDatumHash]))

newtype PTxId (s :: S)

Constructors

PTxId (Term s (PDataRecord '["_0" := PByteString]))

newtype PTxInInfo (s :: S)

A input of the pending transaction.

Constructors

PTxInInfo (Term s (PDataRecord '["outRef" := PTxOutRef, "resolved" := PTxOut]))

Orphan instances

PConstantDecl TxInInfo

Instance details

Associated Types

type PConstantRepr TxInInfo

type PConstanted TxInInfo :: PType

Methods

pconstantToRepr :: TxInInfo -> PConstantRepr TxInInfo

pconstantFromRepr :: PConstantRepr TxInInfo -> Maybe TxInInfo

PConstantDecl TxId

Instance details

Associated Types

type PConstantRepr TxId

type PConstanted TxId :: PType

Methods

pconstantToRepr :: TxId -> PConstantRepr TxId

pconstantFromRepr :: PConstantRepr TxId -> Maybe TxId

PConstantDecl TxOut

Instance details

Associated Types

type PConstantRepr TxOut

type PConstanted TxOut :: PType

Methods

pconstantToRepr :: TxOut -> PConstantRepr TxOut

pconstantFromRepr :: PConstantRepr TxOut -> Maybe TxOut

PConstantDecl TxOutRef

Instance details

Associated Types

type PConstantRepr TxOutRef

type PConstanted TxOutRef :: PType

Methods

pconstantToRepr :: TxOutRef -> PConstantRepr TxOutRef

pconstantFromRepr :: PConstantRepr TxOutRef -> Maybe TxOutRef