Plutarch.Api.V1.Value

Description

Value-related functionality. In order to keep the interface efficient and safe at the same time, there is a type-level distinction between PValues that are guaranteed to be properly normalized and those that provide no such guarantee.

Also for efficiency reasons, the Ada-specific functions assume that there can be only one token name for the Ada currency symbol, and they don't check whether it matches adaToken.

Documentation

newtype PValue (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S)

Constructors

PValue (Term s (PMap keys PCurrencySymbol (PMap keys PTokenName PInteger)))

newtype PCurrencySymbol (s :: S)

Constructors

PCurrencySymbol (Term s PByteString)

newtype PTokenName (s :: S)

Constructors

PTokenName (Term s PByteString)

data KeyGuarantees

Constructors

Sorted

Unsorted

data AmountGuarantees

Constructors

NoGuarantees

NonZero

Positive

Conversions and assertions

passertSorted :: Term s (PValue _ _ :--> PValue 'Sorted 'NonZero)

Assert the value is properly sorted and normalized.

passertPositive :: Term s (PValue 'Sorted 'NonZero :--> PValue 'Sorted 'Positive)

Assert all amounts in the value are positive.

pforgetPositive :: Term s (PValue 'Sorted 'Positive) -> Term s (PValue k a)

Forget the knowledge of value's positivity.

pforgetSorted :: Term s (PValue 'Sorted a) -> Term s (PValue k b)

Forget the knowledge of all value's guarantees.

pnormalize :: Term s (PValue 'Sorted _ :--> PValue 'Sorted 'NonZero)

Normalize the argument to contain no zero quantity nor empty token map.

Creation

psingleton :: Term s (PCurrencySymbol :--> (PTokenName :--> (PInteger :--> PValue 'Sorted 'NonZero)))

Construct a singleton PValue containing only the given quantity of the given currency.

psingletonData :: Term s (PAsData PCurrencySymbol :--> (PAsData PTokenName :--> (PAsData PInteger :--> PValue 'Sorted 'NonZero)))

Construct a singleton PValue containing only the given quantity of the given currency, taking data-encoded parameters.

pconstantSingleton :: ClosedTerm PCurrencySymbol -> ClosedTerm PTokenName -> ClosedTerm PInteger -> ClosedTerm (PValue 'Sorted 'NonZero)

Construct a constant singleton PValue containing only the given quantity of the given currency.

pconstantPositiveSingleton :: ClosedTerm PCurrencySymbol -> ClosedTerm PTokenName -> ClosedTerm PInteger -> ClosedTerm (PValue 'Sorted 'Positive)

Construct a constant singleton PValue containing only the given positive quantity of the given currency.

Combining values

punionWith :: Term s ((PInteger :--> (PInteger :--> PInteger)) :--> (PValue 'Sorted _ :--> (PValue 'Sorted _ :--> PValue 'Sorted 'NoGuarantees)))

Combine two PValues applying the given function to any pair of quantities with the same asset class. Note that the result is _not_ normalized and may contain zero quantities.

punionWithData :: Term s ((PAsData PInteger :--> (PAsData PInteger :--> PAsData PInteger)) :--> (PValue 'Sorted _ :--> (PValue 'Sorted _ :--> PValue 'Sorted 'NoGuarantees)))

Combine two PValues applying the given function to any pair of data-encoded quantities with the same asset class. Note that the result is _not_ normalized and may contain zero quantities.

Lookups

pvalueOf :: Term s (PValue _ _ :--> (PCurrencySymbol :--> (PTokenName :--> PInteger)))

Get the quantity of the given currency in the PValue.

plovelaceValueOf :: Term s (PValue 'Sorted v :--> PInteger)

The amount of Lovelace in value

Ada-specific

padaSymbol :: Term s PCurrencySymbol

The PCurrencySymbol of the Ada currency.

padaSymbolData :: Term s (PAsData PCurrencySymbol)

Data-encoded PCurrencySymbol of the Ada currency.

padaToken :: Term s PTokenName

The PTokenName of the Ada currency.

padaTokenData :: Term s (PAsData PTokenName)

Data-encoded PTokenName of the Ada currency.

pisAdaOnlyValue :: Term s (PValue 'Sorted 'Positive :--> PBool)

Test if the value contains nothing but Ada

padaOnlyValue :: Term s (PValue 'Sorted v :--> PValue 'Sorted v)

Value without any non-Ada

pnoAdaValue :: Term s (PValue 'Sorted v :--> PValue 'Sorted v)

Value without any Ada

Orphan instances

PConstantDecl CurrencySymbol

Instance details

Associated Types

type PConstantRepr CurrencySymbol

type PConstanted CurrencySymbol :: PType

Methods

pconstantToRepr :: CurrencySymbol -> PConstantRepr CurrencySymbol

pconstantFromRepr :: PConstantRepr CurrencySymbol -> Maybe CurrencySymbol

PConstantDecl TokenName

Instance details

Associated Types

type PConstantRepr TokenName

type PConstanted TokenName :: PType

Methods

pconstantToRepr :: TokenName -> PConstantRepr TokenName

pconstantFromRepr :: PConstantRepr TokenName -> Maybe TokenName

PConstantDecl Value

Instance details

Associated Types

type PConstantRepr Value

type PConstanted Value :: PType

Methods

pconstantToRepr :: Value -> PConstantRepr Value

pconstantFromRepr :: PConstantRepr Value -> Maybe Value