Plutarch.Api.V2

Contexts

newtype PScriptContext (s :: S)

Script context consists of the script purpose and the pending transaction info.

Constructors

PScriptContext (Term s (PDataRecord '["txInfo" := PTxInfo, "purpose" := PScriptPurpose]))

newtype PTxInfo (s :: S)

A pending transaction. This is the view as seen by the validator script.

Constructors

PTxInfo (Term s (PDataRecord '["inputs" := PBuiltinList PTxInInfo, "referenceInputs" := PBuiltinList PTxInInfo, "outputs" := PBuiltinList PTxOut, "fee" := PValue 'Sorted 'Positive, "mint" := PValue 'Sorted 'NoGuarantees, "dcert" := PBuiltinList PDCert, "wdrl" := PMap 'Unsorted PStakingCredential PInteger, "validRange" := PPOSIXTimeRange, "signatories" := PBuiltinList (PAsData PPubKeyHash), "redeemers" := PMap 'Unsorted PScriptPurpose PRedeemer, "datums" := PMap 'Unsorted PDatumHash PDatum, "id" := PTxId]))

data PScriptPurpose (s :: S)

The purpose of the script that is currently running

Constructors

PMinting (Term s (PDataRecord '["_0" := PCurrencySymbol]))

PSpending (Term s (PDataRecord '["_0" := PTxOutRef]))

PRewarding (Term s (PDataRecord '["_0" := PStakingCredential]))

PCertifying (Term s (PDataRecord '["_0" := PDCert]))

Tx

newtype PTxOutRef (s :: S)

Reference to a transaction output with a index referencing which of the outputs is being referred to.

Constructors

PTxOutRef (Term s (PDataRecord '["id" := PTxId, "idx" := PInteger]))

newtype PTxOut (s :: S)

A transaction output. This consists of a target address, value and maybe a datum hash

Constructors

PTxOut (Term s (PDataRecord '["address" := PAddress, "value" := PValue 'Sorted 'Positive, "datum" := POutputDatum, "referenceScript" := PMaybeData PScriptHash]))

newtype PTxId (s :: S)

Constructors

PTxId (Term s (PDataRecord '["_0" := PByteString]))

newtype PTxInInfo (s :: S)

A input of the pending transaction.

Constructors

PTxInInfo (Term s (PDataRecord '["outRef" := PTxOutRef, "resolved" := PTxOut]))

data POutputDatum (s :: S)

The datum attached to an output: either nothing, a datum hash or an inline datum (CIP 32)

Constructors

PNoOutputDatum (Term s (PDataRecord '[]))

POutputDatumHash (Term s (PDataRecord '["datumHash" := PDatumHash]))

POutputDatum (Term s (PDataRecord '["outputDatum" := PDatum]))

reexports for unchanged V1 ledger types

data PMaybeData a (s :: S)

Data encoded Maybe type. Used in various ledger api types.

Constructors

PDJust (Term s (PDataRecord '["_0" := a]))

PDNothing (Term s (PDataRecord '[]))

type PTuple a b = PDataSum '['["_0" := a, "_1" := b]]

newtype PDatum (s :: S)

Constructors

PDatum (Term s PData)

newtype PDatumHash (s :: S)

Constructors

PDatumHash (Term s PByteString)

newtype PAddress (s :: S)

Constructors

PAddress (Term s (PDataRecord '["credential" := PCredential, "stakingCredential" := PMaybeData PStakingCredential]))

data KeyGuarantees

Constructors

Sorted

Unsorted

data AmountGuarantees

Constructors

NoGuarantees

NonZero

Positive

newtype PScriptHash (s :: S)

Constructors

PScriptHash (Term s PByteString)

newtype PPubKeyHash (s :: S)

Constructors

PPubKeyHash (Term s PByteString)

data PStakingCredential (s :: S)

Constructors

PStakingHash (Term s (PDataRecord '["_0" := PCredential]))

PStakingPtr (Term s (PDataRecord '["_0" := PInteger, "_1" := PInteger, "_2" := PInteger]))

newtype PValue (keys :: KeyGuarantees) (amounts :: AmountGuarantees) (s :: S)

Constructors

PValue (Term s (PMap keys PCurrencySymbol (PMap keys PTokenName PInteger)))

newtype PCurrencySymbol (s :: S)

Constructors

PCurrencySymbol (Term s PByteString)

newtype PTokenName (s :: S)

Constructors

PTokenName (Term s PByteString)

newtype PInterval a (s :: S)

Constructors

PInterval (Term s (PDataRecord '["from" := PLowerBound a, "to" := PUpperBound a]))

data PExtended a (s :: S)

Constructors

PNegInf (Term s (PDataRecord '[]))

PFinite (Term s (PDataRecord '["_0" := a]))

PPosInf (Term s (PDataRecord '[]))

type PClosure = PBool

newtype PLowerBound a (s :: S)

Constructors

PLowerBound (Term s (PDataRecord '["_0" := PExtended a, "_1" := PClosure]))

newtype PUpperBound a (s :: S)

Constructors

PUpperBound (Term s (PDataRecord '["_0" := PExtended a, "_1" := PClosure]))

newtype PPOSIXTime (s :: S)

Constructors

PPOSIXTime (Term s PInteger)

type PPOSIXTimeRange = PInterval PPOSIXTime

newtype PMap (keysort :: KeyGuarantees) (k :: PType) (v :: PType) (s :: S)

Constructors

PMap (Term s (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))))

Script Utils

scriptHash :: Script -> ScriptHash

Hash a Script, with the correct prefix for Plutus V2

datumHash :: Datum -> DatumHash

Hash a Datum.

redeemerHash :: Redeemer -> RedeemerHash

Hash a Redeemer.

dataHash :: ToData a => a -> BuiltinByteString

Hash the data encoded representation of given argument.

type PValidator = PData :--> (PData :--> (PScriptContext :--> POpaque))

a Validator Term

type PStakeValidator = PData :--> (PScriptContext :--> POpaque)

a StakeValidator Term

type PMintingPolicy = PData :--> (PScriptContext :--> POpaque)

a MintingPolicy Term