Plutarch.Api.V2.Contexts

Documentation

newtype PScriptContext (s :: S)

Script context consists of the script purpose and the pending transaction info.

Constructors

PScriptContext (Term s (PDataRecord '["txInfo" ':= PTxInfo, "purpose" ':= PScriptPurpose]))

newtype PTxInfo (s :: S)

A pending transaction. This is the view as seen by the validator script.

Constructors

PTxInfo (Term s (PDataRecord '["inputs" ':= PBuiltinList PTxInInfo, "referenceInputs" ':= PBuiltinList PTxInInfo, "outputs" ':= PBuiltinList PTxOut, "fee" ':= PValue 'Sorted 'Positive, "mint" ':= PValue 'Sorted 'NoGuarantees, "dcert" ':= PBuiltinList PDCert, "wdrl" ':= PMap 'Unsorted PStakingCredential PInteger, "validRange" ':= PPOSIXTimeRange, "signatories" ':= PBuiltinList (PAsData PPubKeyHash), "redeemers" ':= PMap 'Unsorted PScriptPurpose PRedeemer, "datums" ':= PMap 'Unsorted PDatumHash PDatum, "id" ':= PTxId]))

data PScriptPurpose (s :: S)

The purpose of the script that is currently running

Constructors

PMinting (Term s (PDataRecord '["_0" ':= PCurrencySymbol]))

PSpending (Term s (PDataRecord '["_0" ':= PTxOutRef]))

PRewarding (Term s (PDataRecord '["_0" ':= PStakingCredential]))

PCertifying (Term s (PDataRecord '["_0" ':= PDCert]))

Orphan instances

PConstantDecl ScriptContext

Instance details

Associated Types

type PConstantRepr ScriptContext

type PConstanted ScriptContext :: PType

Methods

pconstantToRepr :: ScriptContext -> PConstantRepr ScriptContext

pconstantFromRepr :: PConstantRepr ScriptContext -> Maybe ScriptContext

PConstantDecl TxInfo

Instance details

Associated Types

type PConstantRepr TxInfo

type PConstanted TxInfo :: PType

Methods

pconstantToRepr :: TxInfo -> PConstantRepr TxInfo

pconstantFromRepr :: PConstantRepr TxInfo -> Maybe TxInfo