Plutarch.Api.V2.Tx

Documentation

newtype PTxOutRef (s :: S)

Reference to a transaction output with a index referencing which of the outputs is being referred to.

Constructors

PTxOutRef (Term s (PDataRecord '["id" := PTxId, "idx" := PInteger]))

newtype PTxOut (s :: S)

A transaction output. This consists of a target address, value and maybe a datum hash

Constructors

PTxOut (Term s (PDataRecord '["address" := PAddress, "value" := PValue 'Sorted 'Positive, "datum" := POutputDatum, "referenceScript" := PMaybeData PScriptHash]))

newtype PTxId (s :: S)

Constructors

PTxId (Term s (PDataRecord '["_0" := PByteString]))

newtype PTxInInfo (s :: S)

A input of the pending transaction.

Constructors

PTxInInfo (Term s (PDataRecord '["outRef" := PTxOutRef, "resolved" := PTxOut]))

data POutputDatum (s :: S)

The datum attached to an output: either nothing, a datum hash or an inline datum (CIP 32)

Constructors

PNoOutputDatum (Term s (PDataRecord '[]))

POutputDatumHash (Term s (PDataRecord '["datumHash" := PDatumHash]))

POutputDatum (Term s (PDataRecord '["outputDatum" := PDatum]))

Orphan instances

PConstantDecl TxInInfo

Instance details

Associated Types

type PConstantRepr TxInInfo

type PConstanted TxInInfo :: PType

Methods

pconstantToRepr :: TxInInfo -> PConstantRepr TxInInfo

pconstantFromRepr :: PConstantRepr TxInInfo -> Maybe TxInInfo

PConstantDecl OutputDatum

Instance details

Associated Types

type PConstantRepr OutputDatum

type PConstanted OutputDatum :: PType

Methods

pconstantToRepr :: OutputDatum -> PConstantRepr OutputDatum

pconstantFromRepr :: PConstantRepr OutputDatum -> Maybe OutputDatum

PConstantDecl TxOut

Instance details

Associated Types

type PConstantRepr TxOut

type PConstanted TxOut :: PType

Methods

pconstantToRepr :: TxOut -> PConstantRepr TxOut

pconstantFromRepr :: PConstantRepr TxOut -> Maybe TxOut