Plutarch.BLS

Documentation

newtype PBuiltinBLS12_381_G1_Element s

A point on the BLS12-381 G1 curve.

Since: 1.9.0

Constructors

PBuiltinBLS12_381_G1_Element (Term s POpaque)

pbls12_381_G1_add :: Term s (PBuiltinBLS12_381_G1_Element :--> (PBuiltinBLS12_381_G1_Element :--> PBuiltinBLS12_381_G1_Element))

Add two points on the BLS12-381 G1 curve.

Since: 1.9.0

pbls12_381_G1_scalarMul :: Term s (PInteger :--> (PBuiltinBLS12_381_G1_Element :--> PBuiltinBLS12_381_G1_Element))

Multiply a point on the BLS12-381 G1 curve by a scalar.

Since: 1.9.0

pbls12_381_G1_neg :: Term s (PBuiltinBLS12_381_G1_Element :--> PBuiltinBLS12_381_G1_Element)

Negate a point on the BLS12-381 G1 curve.

Since: 1.9.0

pbls12_381_G1_compress :: Term s (PBuiltinBLS12_381_G1_Element :--> PByteString)

Compress a point on the BLS12-381 G1 curve to a byte string.

Since: 1.9.0

pbls12_381_G1_uncompress :: Term s (PByteString :--> PBuiltinBLS12_381_G1_Element)

Uncompress a byte string to a point on the BLS12-381 G1 curve.

Since: 1.9.0

pbls12_381_G1_hashToGroup :: Term s (PByteString :--> (PByteString :--> PBuiltinBLS12_381_G1_Element))

Hash a message to a point on the BLS12-381 G1 curve.

Since: 1.9.0

pbls12_381_G1_compressed_zero :: Term s PByteString

The compressed representation of the zero point on the BLS12-381 G1 curve.

Since: 1.9.0

pbls12_381_G1_compressed_generator :: Term s PByteString

The compressed representation of the generator point on the BLS12-381 G1 curve.

Since: 1.9.0

newtype PBuiltinBLS12_381_G2_Element s

Since: 1.9.0

Constructors

PBuiltinBLS12_381_G2_Element (Term s POpaque)

pbls12_381_G2_add :: Term s (PBuiltinBLS12_381_G2_Element :--> (PBuiltinBLS12_381_G2_Element :--> PBuiltinBLS12_381_G2_Element))

Add two points on the BLS12-381 G2 curve.

Since: 1.9.0

pbls12_381_G2_scalarMul :: Term s (PInteger :--> (PBuiltinBLS12_381_G2_Element :--> PBuiltinBLS12_381_G2_Element))

Multiply a point on the BLS12-381 G2 curve by a scalar.

Since: 1.9.0

pbls12_381_G2_neg :: Term s (PBuiltinBLS12_381_G2_Element :--> PBuiltinBLS12_381_G2_Element)

Negate a point on the BLS12-381 G2 curve.

Since: 1.9.0

pbls12_381_G2_compress :: Term s (PBuiltinBLS12_381_G2_Element :--> PByteString)

Compress a point on the BLS12-381 G2 curve to a byte string.

Since: 1.9.0

pbls12_381_G2_uncompress :: Term s (PByteString :--> PBuiltinBLS12_381_G2_Element)

Uncompress a byte string to a point on the BLS12-381 G2 curve.

Since: 1.9.0

pbls12_381_G2_hashToGroup :: Term s (PByteString :--> (PByteString :--> PBuiltinBLS12_381_G2_Element))

Hash a message to a point on the BLS12-381 G2 curve.

Since: 1.9.0

pbls12_381_G2_compressed_zero :: Term s PByteString

The compressed representation of the zero point on the BLS12-381 G2 curve.

Since: 1.9.0

pbls12_381_G2_compressed_generator :: Term s PByteString

The compressed representation of the generator point on the BLS12-381 G2 curve.

Since: 1.9.0

newtype PBuiltinBLS12_381_MlResult s

Represents the result of a Miller loop operation in BLS12-381 pairing.

Since: 1.9.0

Constructors

PBuiltinBLS12_381_MlResult (Term s POpaque)

pbls12_381_millerLoop :: Term s (PBuiltinBLS12_381_G1_Element :--> (PBuiltinBLS12_381_G2_Element :--> PBuiltinBLS12_381_MlResult))

Perform a Miller loop operation on a G1 and G2 element.

Since: 1.9.0

pbls12_381_mulMlResult :: Term s (PBuiltinBLS12_381_MlResult :--> (PBuiltinBLS12_381_MlResult :--> PBuiltinBLS12_381_MlResult))

Multiply two Miller loop results.

Since: 1.9.0

pbls12_381_finalVerify :: Term s (PBuiltinBLS12_381_MlResult :--> (PBuiltinBLS12_381_MlResult :--> PBool))

Perform the final verification step in BLS12-381 pairing.

Since: 1.9.0