Plutarch.Bool

Documentation

data PBool (s :: S)

Plutus BuiltinBool

Constructors

PTrue

PFalse

class PEq t where

Minimal complete definition

Nothing

Methods

(#==) :: Term s t -> Term s t -> Term s PBool infix 4

default (#==) :: (PGeneric t, PlutusType t, All2 PEq (PCode t)) => Term s t -> Term s t -> Term s PBool

class PEq t => PPartialOrd t where

Partial ordering relation.

Laws

#<= must form a partial order. More precisely:

x #<= x = True (#<= is reflexive)

(x #<= y) #&& (y #<= x) = x #== y (#<= is anti-symmetric)

(x #<= y) #&& (y #<= z) = x #<= z (@#<= is transitive)

Furthermore, #< must be an equivalent strict partial order to #<=:

x #<= y = (x #< y) #|| (x #== y)

x #< x = False (#< is irreflexive)

x #< y = pnot (y #< x) (#< is asymmetric)

(x #< y) #&& (y #< z) = x #< z (#< is transitive)

Lastly, if you define #>= or #>, ensure that the following also hold:

x #> y = y #< x

x #>= y = pnot (x #< y)

The default implementations of #>= and #> ensure these laws.

Minimal complete definition

Nothing

Methods

(#<=) :: Term s t -> Term s t -> Term s PBool infix 4

default (#<=) :: POrd (PInner t) => Term s t -> Term s t -> Term s PBool

(#<) :: Term s t -> Term s t -> Term s PBool infix 4

default (#<) :: POrd (PInner t) => Term s t -> Term s t -> Term s PBool

(#>=) :: forall (s :: S). Term s t -> Term s t -> Term s PBool infix 4

@since WIP

(#>) :: forall (s :: S). Term s t -> Term s t -> Term s PBool infix 4

@since WIP

class PPartialOrd t => POrd t where

Total ordering relation.

Laws

pmax and pmin must form a commutative semiring without identity elements, where addition also distributes over multiplication. More precisely:

pmax x y = pmax y x (pmax is commutative)

pmin x y = pmin y x (pmin is commutative)

pmax x (pmax y z) = pmax (pmax x y) z (pmax is associative)

pmin x (pmin y z) = pmin (pmin x y) z (pmin is associative)

pmax x (pmin y z) = pmin (pmax x y) (pmax x z) (pmax distributes over pmin)

pmin x (pmax y z) = pmax (pmin x y) (pmin x z) (pmin distributes over pmax)

Furthermore, the following must hold relative #<:

pmin x y = if (x #< y) then x else y

pmax x y = if (x #< y) then y else x

Laws 7 and 8 are also the defaults, as for most types, this is the best you can do.

Minimal complete definition

Nothing

Methods

pmax :: forall (s :: S). Term s t -> Term s t -> Term s t

@since WIP

pmin :: forall (s :: S). Term s t -> Term s t -> Term s t

@since WIP

pif :: Term s PBool -> Term s a -> Term s a -> Term s a

Lazy if-then-else.

pif' :: Term s (PBool :--> (a :--> (a :--> a)))

Strict version of pif. Emits slightly less code.

pnot :: Term s (PBool :--> PBool)

Boolean negation for PBool terms.

(#&&) :: Term s PBool -> Term s PBool -> Term s PBool infixr 3

Lazily evaluated boolean and for PBool terms.

(#||) :: Term s PBool -> Term s PBool -> Term s PBool infixr 2

Lazily evaluated boolean or for PBool terms.

por :: Term s (PBool :--> (PDelayed PBool :--> PDelayed PBool))

Hoisted, Plutarch level, lazily evaluated boolean or function.

pand :: Term s (PBool :--> (PDelayed PBool :--> PDelayed PBool))

Hoisted, Plutarch level, lazily evaluated boolean and function.

pand' :: Term s (PBool :--> (PBool :--> PBool))

Hoisted, Plutarch level, strictly evaluated boolean and function.

por' :: Term s (PBool :--> (PBool :--> PBool))

Hoisted, Plutarch level, strictly evaluated boolean or function.

data PSBool (s :: S)

Scott-encoded bool.

Constructors

PSTrue

PSFalse

pmatchStrict :: forall (r :: PType) (s :: S). Term s PSBool -> (PSBool s -> Term s r) -> Term s r

Strict version of pmatch for PSBool.

pstrue :: forall (s :: S). Term s PSBool

psfalse :: forall (s :: S). Term s PSBool

psif :: forall (s :: S) (a :: PType). Term s PSBool -> Term s a -> Term s a -> Term s a

Lazy if on Scott-encoded bool.

psif' :: forall (s :: S) (a :: PType). Term s PSBool -> Term s a -> Term s a -> Term s a

Strict if on Scott-encoded bool.

psnot :: forall (s :: S). Term s PSBool -> Term s PSBool

not on Scott-encoded bool.

psand :: forall (s :: S). Term s PSBool -> Term s PSBool -> Term s PSBool

psand' :: forall (s :: S). Term s PSBool -> Term s PSBool -> Term s PSBool

psor :: forall (s :: S). Term s PSBool -> Term s PSBool -> Term s PSBool

psor' :: forall (s :: S). Term s PSBool -> Term s PSBool -> Term s PSBool

Orphan instances

PConstantDecl Bool

Instance details

Associated Types

type PConstantRepr Bool

type PConstanted Bool :: PType

Methods

pconstantToRepr :: Bool -> PConstantRepr Bool

pconstantFromRepr :: PConstantRepr Bool -> Maybe Bool