Plutarch.Builtin

Documentation

data PData (s :: S)

pfstBuiltin :: Term s (PBuiltinPair a b :--> a)

psndBuiltin :: Term s (PBuiltinPair a b :--> b)

pasConstr :: Term s (PData :--> PBuiltinPair PInteger (PBuiltinList PData))

pasMap :: Term s (PData :--> PBuiltinList (PBuiltinPair PData PData))

pasList :: Term s (PData :--> PBuiltinList PData)

pasInt :: Term s (PData :--> PInteger)

plistData :: Term s (PBuiltinList PData :--> PData)

pconstantData :: forall p h s. (ToData h, PLifted p ~ h, PConstanted h ~ p) => h -> Term s (PAsData p)

Create a Plutarch-level PAsData constant, from a Haskell value. Example: > pconstantData @PInteger 42

pconstrBuiltin :: Term s (PInteger :--> (PBuiltinList PData :--> PAsData (PBuiltinPair PInteger (PBuiltinList PData))))

pasByteStr :: Term s (PData :--> PByteString)

data PBuiltinPair (a :: PType) (b :: PType) (s :: S)

Plutus BuiltinPair

data PBuiltinList (a :: PType) (s :: S)

Plutus BuiltinList

Constructors

PCons (Term s a) (Term s (PBuiltinList a))

PNil

pdataLiteral :: Data -> Term s PData

Deprecated: Use pconstant instead.

class PIsData a where

Laws: - If PSubtype PData a, then pdataImpl a must be pupcast. - pdataImpl . pupcast . pfromDataImpl ≡ id - pfromDataImpl . punsafeDowncast . pdataImpl ≡ id

Minimal complete definition

Nothing

Methods

pfromDataImpl :: Term s (PAsData a) -> Term s a

default pfromDataImpl :: PIsData (PInner a) => Term s (PAsData a) -> Term s a

pdataImpl :: Term s a -> Term s PData

default pdataImpl :: PIsData (PInner a) => Term s a -> Term s PData

pdata :: PIsData a => Term s a -> Term s (PAsData a)

pfromData :: PIsData a => Term s (PAsData a) -> Term s a

data PAsData (a :: PType) (s :: S)

pforgetData :: forall s a. Term s (PAsData a) -> Term s PData

prememberData :: forall (p :: PType -> PType) s. PVariant p => Proxy p -> Term s (p PData) -> Term s (p (PAsData PData))

Inverse of pforgetData'.

prememberData' :: forall a (p :: PType -> PType) s. (PSubtype PData a, PVariant p) => Proxy p -> Term s (p a) -> Term s (p (PAsData a))

Like prememberData but generalised.

pserialiseData :: Term s (PData :--> PByteString)

Serialise any builtin data to its cbor represented by a builtin bytestring

ppairDataBuiltin :: Term s (PAsData a :--> (PAsData b :--> PBuiltinPair (PAsData a) (PAsData b)))

Construct a builtin pair of PData elements.

Uses PAsData to preserve more information about the underlying PData.

pchooseListBuiltin :: Term s (PBuiltinList a :--> (b :--> (b :--> b)))

pchooseData :: Term s (PData :--> (a :--> (a :--> (a :--> (a :--> (a :--> a))))))

Orphan instances

PConstantDecl Data

Instance details

Associated Types

type PConstantRepr Data

type PConstanted Data :: PType

Methods

pconstantToRepr :: Data -> PConstantRepr Data

pconstantFromRepr :: PConstantRepr Data -> Maybe Data

PConstant a => PConstantDecl [a]

Instance details

Associated Types

type PConstantRepr [a]

type PConstanted [a] :: PType

Methods

pconstantToRepr :: [a] -> PConstantRepr [a]

pconstantFromRepr :: PConstantRepr [a] -> Maybe [a]

(PConstant a, PConstant b) => PConstantDecl (a, b)

Instance details

Associated Types

type PConstantRepr (a, b)

type PConstanted (a, b) :: PType

Methods

pconstantToRepr :: (a, b) -> PConstantRepr (a, b)

pconstantFromRepr :: PConstantRepr (a, b) -> Maybe (a, b)