Plutarch.Builtin.Data

Documentation

newtype PData (s :: S)

Constructors

PData (Term s PData)

newtype PAsData (a :: S -> Type) (s :: S)

Constructors

PAsData (Term s a)

pchooseData :: Term s (PData :--> (a :--> (a :--> (a :--> (a :--> (a :--> a))))))

pasConstr :: Term s (PData :--> PBuiltinPair PInteger (PBuiltinList PData))

pasMap :: Term s (PData :--> PBuiltinList (PBuiltinPair PData PData))

plistData :: Term s (PBuiltinList PData :--> PData)

pasList :: Term s (PData :--> PBuiltinList PData)

pasInt :: Term s (PData :--> PInteger)

pasByteStr :: Term s (PData :--> PByteString)

pserialiseData :: Term s (PData :--> PByteString)

Serialise any builtin data to its cbor represented by a builtin bytestring

pconstrBuiltin :: Term s (PInteger :--> (PBuiltinList PData :--> PAsData (PBuiltinPair PInteger (PBuiltinList PData))))

newtype PBuiltinPair (a :: S -> Type) (b :: S -> Type) (s :: S)

Constructors

PBuiltinPair (Term s (PBuiltinPair a b))

pfstBuiltin :: Term s (PBuiltinPair a b :--> a)

psndBuiltin :: Term s (PBuiltinPair a b :--> b)

ppairDataBuiltin :: Term s (PAsData a :--> (PAsData b :--> PBuiltinPair (PAsData a) (PAsData b)))

Construct a builtin pair of PData elements.

Uses PAsData to preserve more information about the underlying PData.

data PBuiltinList (a :: S -> Type) (s :: S)

Plutus BuiltinList

Constructors

PCons (Term s a) (Term s (PBuiltinList a))

PNil

pheadBuiltin :: Term s (PBuiltinList a :--> a)

ptailBuiltin :: Term s (PBuiltinList a :--> PBuiltinList a)

pchooseListBuiltin :: Term s (PBuiltinList a :--> (b :--> (b :--> b)))

pnullBuiltin :: Term s (PBuiltinList a :--> PBool)

pconsBuiltin :: Term s (a :--> (PBuiltinList a :--> PBuiltinList a))