Plutarch.Builtin.Integer

Type

data PInteger s

A builtin Plutus integer.

@since WIP

Functions

pexpModInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> (PInteger :--> PInteger)))

Performs modulo exponentiation. More precisely, pexpModInteger b e m performs b to the power of e, modulo m. The result is always non-negative.

Note

This will error if the modulus is zero. When given a negative exponent, this will try to find a modular multiplicative inverse, and will error if none exists.

@since WIP

peqInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PBool))

pleInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PBool))

pltInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PBool))

paddInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

psubtractInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

pmultiplyInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

pconstantInteger :: forall (s :: S). Integer -> Term s PInteger

pquotientInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

premainderInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))