Plutarch.ByteString

Types

data PByte (s :: S)

A Plutarch-level representation of bytes.

Note =

This type is intentionally quite restrictive, as it's not really meant to be computed with. Instead, it ensures certain operations' type safety while also allowing more sensible signatures. If you want to do anything with PBytes, we recommend converting them to PIntegers first.

@since WIP

data PByteString s

Plutus BuiltinByteString

data PLogicOpSemantics (s :: S)

Type designating whether logical operations should use padding or truncation semantics. See CIP-122 for more details on this.

@since WIP

Functions

Construction

ppadding :: forall (s :: S). Term s PLogicOpSemantics

Indicates that padding semantics should be used.

@since WIP

ptruncation :: forall (s :: S). Term s PLogicOpSemantics

Indicates that truncation semantics should be used.

@since WIP

pzeroesBS :: forall (s :: S). Term s (PInteger :--> PByteString)

Construct a PByteString of the specified length (0 if negative) consisting entirely of zero bytes.

@since WIP

ponesBS :: forall (s :: S). Term s (PInteger :--> PByteString)

Construct a PByteString of the specified length (0 if negative) consisting entirely of ones; that is, where every byte is 0xFF.

@since WIP

preplicateBS :: forall (s :: S). Term s (PInteger :--> (PByte :--> PByteString))

Given a desired length and a PByte, construct a PByteString of the specified length (0 if negative) consisting entirely of that PByte.

@since WIP

Byte-oriented

pallBS :: forall (s :: S). Term s ((PByte :--> PBool) :--> (PByteString :--> PBool))

Verify that the given predicate holds for every byte in the argument.

@since WIP

pandBS :: forall (s :: S). Term s (PLogicOpSemantics :--> (PByteString :--> (PByteString :--> PByteString)))

Perform the logical AND of two PByteStrings, as per CIP-122. The PLogicOpSemantics argument specifies what should be done if the lengths of the two PByteString arguments do not match.

@since WIP

porBS :: forall (s :: S). Term s (PLogicOpSemantics :--> (PByteString :--> (PByteString :--> PByteString)))

Perform the logical OR of two PByteStrings, as per CIP-122. The PLogicOpSemantics argument specifies what should be done if the lengths of the two PByteString arguments do not match.

@since WIP

pxorBS :: forall (s :: S). Term s (PLogicOpSemantics :--> (PByteString :--> (PByteString :--> PByteString)))

Perform the logical XOR of two PByteStrings, as per CIP-122. The PLogicOpSemantics argument specifies what should be done if the lengths of the two PByteString arguments do not match.

@since WIP

pcomplementBS :: forall (s :: S). Term s (PByteString :--> PByteString)

Perform the logical complement of a PByteString, as per CIP-122.

@since WIP

pconsBS :: Term s (PByte :--> (PByteString :--> PByteString))

Prepend a PByte to a 'PByteString.

@since WIP

psliceBS :: Term s (PInteger :--> (PInteger :--> (PByteString :--> PByteString)))

Slice a PByteString with given start index and slice length.

>>> (pslice # 2 # 3 phexByteStr "4102afde5b2a") #== phexByteStr "afde5b"

plengthBS :: Term s (PByteString :--> PInteger)

Find the length of a PByteString.

pindexBS :: Term s (PByteString :--> (PInteger :--> PByte))

Given a valid index into a PByteString, returns the PByte at that index. Will crash if given an out-of-bounds index.

@since WIP

Conversion

pbyteToInteger :: Term s (PByte :--> PInteger)

Convert a PByte into its corresponding PInteger.

@since WIP

pintegerToByte :: Term s (PInteger :--> PByte)

Try to convert a PInteger into its corresponding PByte. This operation is checked, and will error if given a negative PInteger, or one too large to fit into a byte.

@since WIP

punsafeIntegerToByte :: Term s (PInteger :--> PByte)

As pintegerToByte, but unchecked. Be very careful with this: it can potentially break downstream invariants and wreak all manner of hard-to-discover havoc. Only included for reasons of efficiency.

@since WIP

Other

phexByteStr :: HasCallStack => String -> Term s PByteString

Interpret a hex string as a PByteString.

Orphan instances

PConstantDecl Word8
@since WIP

Instance details

Associated Types

type PConstantRepr Word8

type PConstanted Word8 :: PType

Methods

pconstantToRepr :: Word8 -> PConstantRepr Word8

pconstantFromRepr :: PConstantRepr Word8 -> Maybe Word8

PConstantDecl ByteString

Instance details

Associated Types

type PConstantRepr ByteString

type PConstanted ByteString :: PType

Methods

pconstantToRepr :: ByteString -> PConstantRepr ByteString

pconstantFromRepr :: PConstantRepr ByteString -> Maybe ByteString