Plutarch.DataRepr.Internal

Documentation

newtype PDataSum defs s

A sum of PDataRecords. The underlying representation is the Constr constructor, where the integer is the index of the variant and the list is the record.

Constructors

PDataSum (NS (Compose (Term s) PDataRecord) defs)

punDataSum :: Term s (PDataSum '[def] :--> PDataRecord def)

If there is only a single variant, then we can safely extract it.

ptryIndexDataSum :: KnownNat n => Proxy n -> Term s (PDataSum (def ': defs) :--> PDataRecord (IndexList n (def ': defs)))

Try getting the nth variant. Errs if it's another variant.

pdcons :: forall label a l s. Term s (PAsData a :--> (PDataRecord l :--> PDataRecord ((label ':= a) ': l)))

Cons a field to a data record.

You can specify the label to associate with the field using type applications-

foo :: Term s (PDataRecord '[ "fooField" ':= PByteString ])
foo = pdcons @"fooField" # pdata (phexByteStr "ab") # pdnil

pdnil :: Term s (PDataRecord '[])

An empty PDataRecord.

data DataReprHandlers (out :: S -> Type) (defs :: [[PLabeledType]]) (s :: S) where

Constructors

DRHNil :: DataReprHandlers out '[] s

DRHCons :: (Term s (PDataRecord def) -> Term s out) -> DataReprHandlers out defs s -> DataReprHandlers out (def ': defs) s

data PDataRecord (as :: [PLabeledType]) (s :: S) where

A "record" of `exists a. PAsData a`. The underlying representation is `PBuiltinList PData`.

Constructors

PDCons :: forall name_x x xs s. PUnLabel name_x ~ x => Term s (PAsData x) -> Term s (PDataRecord xs) -> PDataRecord (name_x ': xs) s

PDNil :: PDataRecord '[] s

data PLabeledType

Constructors

Symbol := (S -> Type)

type family PLabelIndex (name :: Symbol) (as :: [PLabeledType]) :: Nat where ...

Equations

PLabelIndex name ((name ':= _) ': _) = 0

PLabelIndex name (_ ': as) = PLabelIndex name as + 1

type family PUnLabel (a :: PLabeledType) :: S -> Type where ...

Equations

PUnLabel (_ ':= a) = a

type family PLookupLabel name as where ...

Equations

PLookupLabel name ((name ':= a) ': _) = a

PLookupLabel name (_ ': as) = PLookupLabel name as

pindexDataRecord :: KnownNat n => Proxy n -> Term s (PDataRecord as) -> Term s (PAsData (PUnLabel (IndexList n as)))

Safely index a PDataRecord.

pdropDataRecord :: KnownNat n => Proxy n -> Term s (PDataRecord xs) -> Term s (PDataRecord (Drop n xs))

Safely drop the first n items of a PDataRecord.

newtype DualReprHandler s out def

Constructors

DualRepr (Term s (PDataRecord def) -> Term s (PDataRecord def) -> Term s out)

data PlutusTypeData