Plutarch.FFI

Documentation

type family (p :: PType) >~< (t :: Type) :: Constraint where ...

Equality of inner types - Plutarch on the left and Haskell on the right.

Equations

PBool >~< BuiltinBool = ()

PInteger >~< Integer = ()

PString >~< BuiltinString = ()

PByteString >~< BuiltinByteString = ()

PData >~< BuiltinData = ()

PUnit >~< BuiltinUnit = ()

(a :--> b) >~< (a' -> b') = (a >~< a', b >~< b')

(PTxList a) >~< [a'] = a >~< a'

(PTxMaybe a) >~< (Maybe a') = a >~< a'

(PDelayed p) >~< t = (DPTStrat p ~ PlutusTypeScott, G (PCode p) (TypeEncoding t))

data PTxList (a :: PType) (s :: S)

Plutarch type of lists compatible with the PlutusTx encoding of Haskell lists and convertible with the regular PList using plistToTx and plistFromTx.

Constructors

PTxCons (Term s a) (Term s (PTxList a))

PTxNil

data PTxMaybe (a :: PType) (s :: S)

Plutarch type compatible with the PlutusTx encoding of Haskell Maybe and convertible with the regular PMaybe using pmaybeToTx and pmaybeFromTx.

Constructors

PTxJust (Term s a)

PTxNothing

foreignExport :: forall p t. p >~< t => Config -> ClosedTerm p -> CompiledCode t

Compile and export a Plutarch term so it can be used by applyCode.

foreignImport :: forall p t. p >~< t => CompiledCode t -> ClosedTerm p

Import compiled UPLC code (such as a spliced compile result) as a Plutarch term.

opaqueExport :: Config -> ClosedTerm p -> CompiledCode Void

Export Plutarch term of any type as CompiledCode Void.

opaqueImport :: CompiledCode t -> ClosedTerm POpaque

Import compiled UPLC code of any type as a Plutarch opaque term.

plistFromTx :: Term s (PTxList a :--> PList a)

Convert a PTxList to a PList, probably after importing it with foreignImport.

plistToTx :: Term s (PList a :--> PTxList a)

Convert a PList to a PTxList, perhaps before exporting it with foreignExport.

pmaybeFromTx :: Term s (PTxMaybe a :--> PMaybe a)

Convert a PTxMaybe to a PMaybe, probably after importing it with foreignImport.

pmaybeToTx :: Term s (PMaybe a :--> PTxMaybe a)

Convert a PMaybe to a PTxMaybe, perhaps before exporting it with foreignExport.

unsafeForeignExport :: Config -> ClosedTerm p -> CompiledCode t

Seriously unsafe, may fail at run time or result in unexpected behaviour in your on-chain validator.

unsafeForeignImport :: CompiledCode t -> ClosedTerm p

Seriously unsafe, may fail at run time or result in unexpected behaviour in your on-chain validator.