Plutarch.Integer

Type

data PInteger s

Plutus BuiltinInteger

Type class

class PIntegral a where

Minimal complete definition

Nothing

Methods

pdiv :: Term s (a :--> (a :--> a))

default pdiv :: PIntegral (PInner a) => Term s (a :--> (a :--> a))

pmod :: Term s (a :--> (a :--> a))

default pmod :: PIntegral (PInner a) => Term s (a :--> (a :--> a))

pquot :: Term s (a :--> (a :--> a))

default pquot :: PIntegral (PInner a) => Term s (a :--> (a :--> a))

prem :: Term s (a :--> (a :--> a))

default prem :: PIntegral (PInner a) => Term s (a :--> (a :--> a))

Functions

pexpModInteger :: forall (s :: S). Term s (PInteger :--> (PInteger :--> (PInteger :--> PInteger)))

Performs modulo exponentiation. More precisely, pexpModInteger b e m performs b to the power of e, modulo m. The result is always non-negative.

Note

This will error if the modulus is zero. When given a negative exponent, this will try to find a modular multiplicative inverse, and will error if none exists.

@since WIP

Orphan instances

PConstantDecl Integer

Instance details

Associated Types

type PConstantRepr Integer

type PConstanted Integer :: PType

Methods

pconstantToRepr :: Integer -> PConstantRepr Integer

pconstantFromRepr :: PConstantRepr Integer -> Maybe Integer