Plutarch.Internal.Evaluate

Documentation

uplcVersion :: Version

evalScript :: Script -> (Either EvalError Script, ExBudget, [Text])

Evaluate a script with a big budget, returning the trace log and term result.

evalScriptHuge :: Script -> (Either EvalError Script, ExBudget, [Text])

Evaluate a script with a huge budget, returning the trace log and term result.

evalScriptUnlimited :: Script -> (Either (CekEvaluationException NamedDeBruijn DefaultUni DefaultFun) Script, ExBudget, [Text])

Evaluate a script without budget limit

@since WIP

evalScript' :: ExBudget -> Script -> (Either (CekEvaluationException NamedDeBruijn DefaultUni DefaultFun) Script, ExBudget, [Text])

Evaluate a script with a specific budget, returning the trace log and term result.

type EvalError = CekEvaluationException NamedDeBruijn DefaultUni DefaultFun