Plutarch.Internal.Lift

Type class

class PlutusType a => PLiftable (a :: S -> Type) where

Indicates that the given Plutarch type has an equivalent in Haskell (and Plutus by extension), and we have the ability to move between them.

Important note

Calling toPlutarch and fromPlutarch directly should rarely, if ever, be a thing you do. Prefer using pconstant and plift, as these handle some of the oddities required without you having to think about them.

You should rarely, if ever, need to define PLiftable instances by hand. Whenever possible, prefer using DeriveBuiltinPLiftable and DeriveDataPLiftable, as they have fewer complexities and caveats. See their documentation for when to use them.

If you do want to define the methods yourself, there's a few key factors to keep in mind:

PLifted is a trivial wrapper to assist with via derivations: it is isomorphic to Identity and should be treated as such. Use unPLifted to 'pull' terms out.

fromPlutarch involves compilation and evaluation. When doing so, ensure that you use the largest possible budgets to avoid weird problems.

When choosing a type for AsHaskell, any value of that type must be representable in Plutarch. If you have internal invariants to maintain on the Haskell side, make sure you do so with great care.

Laws

fromPlutarch . toPlutarch = Right

fmap toPlutarch . fromPlutarch = Right

Any derivations via DeriveBuiltinPLiftable and DeriveDataPLiftable automatically follow these laws.

@since WIP

Associated Types

type AsHaskell a :: Type

Methods

toPlutarch :: forall (s :: S). AsHaskell a -> PLifted a s

fromPlutarch :: (forall (s :: S). PLifted a s) -> Either LiftError (AsHaskell a)

Functions

pconstant :: forall (a :: S -> Type) (s :: S). PLiftable a => AsHaskell a -> Term s a

Given a Haskell-level representation of a Plutarch term, transform it into its equivalent term.

@since WIP

plift :: forall (a :: S -> Type). PLiftable a => (forall (s :: S). Term s a) -> AsHaskell a

Given a closed Plutarch term, compile and evaluate it, then produce the corresponding Haskell value. If compilation or evaluation fails somehow, this will call error: if you need to 'trap' these outcomes and handle them differently somehow, use fromPlutarch.

@since WIP

Derivation

Via-helpers

newtype DeriveBuiltinPLiftable (a :: S -> Type) (h :: Type) (s :: S)

via-deriving helper, indicating that a has a Haskell-level equivalent h that is directly part of the Plutus default universe (instead of by way of a Data encoding).

@since WIP

Constructors

DeriveBuiltinPLiftable (a s)

newtype DeriveDataPLiftable (a :: S -> Type) (h :: Type) (s :: S)

via-deriving helper, indicating that a has a Haskell-level equivalent h by way of its Data encoding, rather than by h being directly part of the Plutus default universe.

@since WIP

Constructors

DeriveDataPLiftable (a s)

Manual instance helpers

newtype PLifted a s

Similar to Identity, but at the level of Plutarch. Only needed when writing manual instances of PLiftable, or if you want to use toPlutarch and fromPlutarch directly.

@since WIP

Constructors

PLifted (Term s POpaque)

unPLifted :: PLifted a s -> Term s a

Wrapper around punsafeCoerce to 'pull out' a term inside a PLifted.

@since WIP

data LiftError

Used with fromPlutarch methods to give additional information about why evaluating a Plutarch term into a Haskell value went wrong.

@since WIP

Constructors

CouldNotEvaluate EvalError
Evaluation failed for some reason.

TypeError BuiltinError
We tried to use a builtin not part of the Plutus universe.

CouldNotCompile Text
Compiling the term into a script failed.

CouldNotDecodeData
Data encoding was invalid for our type

unsafeToUni :: forall (a :: S -> Type) (h :: Type). DefaultUni `Includes` h => (forall (s :: S). PLifted a s) -> Either LiftError h

Helper for writing PLiftable instances. This is highly unsafe, and only suitable for internal use!

@since WIP

unsafeFromUni :: forall (a :: S -> Type) (h :: Type) (s :: S). DefaultUni `Includes` h => h -> PLifted a s

Helper for writing PLifted instances. This is highly unsafe, and only suitable for internal use!

@since WIP