Plutarch.Internal.Numeric

Types

data PPositive (s :: S)

Since: 1.10.0

data Positive

Since: 1.10.0

data PNatural (s :: S)

Since: 1.10.0

Type classes

class PAdditiveSemigroup (a :: S -> Type) where

The addition operation, and the notion of scaling by a positive.

Laws

x #+ y = y #+ x (commutativity of #+)

x #+ (y #+ z) = (x #+ y) #+ z (associativity of #+)

If you define a custom pscalePositive, ensure the following also hold:

pscalePositive x pone = x

(pscalePositive x n) #+ (pscalePositive x m) = pscalePositive x (n #+ m)

pscalePositive (pscalePositive x n) m = pscalePositive x (n #* m)

The default implementation ensures these laws are satisfied.

Since: 1.10.0

Minimal complete definition

Nothing

Methods

(#+) :: forall (s :: S). Term s a -> Term s a -> Term s a infix 6

default (#+) :: forall (s :: S). PAdditiveSemigroup (PInner a) => Term s a -> Term s a -> Term s a

pscalePositive :: forall (s :: S). Term s a -> Term s PPositive -> Term s a

This defaults to exponentiation-by-squaring, which in general is the best we can do.

class PAdditiveSemigroup a => PAdditiveMonoid (a :: S -> Type) where

The notion of zero, as well as a way to scale by naturals.

Laws

pzero #+ x = x (pzero is the identity of #+)

pscalePositive pzero n = pzero (pzero does not scale up)

If you define pscaleNatural, ensure the following as well:

pscaleNatural x (ppositiveToNatural # p) = pscalePositive x p

pscaleNatural x pzero = pzero

The default implementation of pscaleNatural ensures these laws hold.

Since: 1.10.0

Minimal complete definition

pzero

Methods

pzero :: forall (s :: S). Term s a

pscaleNatural :: forall (s :: S). Term s a -> Term s PNatural -> Term s a

class PAdditiveMonoid a => PAdditiveGroup (a :: S -> Type) where

The notion of additive inverses, and the subtraction operation.

Laws

If you define pnegate, the following laws must hold:

(pnegate # x) #+ x = pzero (pnegate is an additive inverse)

pnegate #$ pnegate # x = x (pnegate is self-inverting)

If you define #-, the following law must hold:

x #- x = pzero

Additionally, the following 'consistency laws' must hold. Default implementations of both pnegate and #- uphold these.

pnegate # x = pzero #- x

x #- y = x #+ (pnegate # y)

Lastly, if you define a custom pscaleInteger, the following laws must hold:

pscaleInteger x pzero = pzero

pscaleInteger x (pnegate # y) = pnegate # (pscaleInteger x y)

Since: 1.10.0

Minimal complete definition

pnegate | (#-)

Methods

pnegate :: forall (s :: S). Term s (a :--> a)

(#-) :: forall (s :: S). Term s a -> Term s a -> Term s a infix 6

pscaleInteger :: forall (s :: S). Term s a -> Term s PInteger -> Term s a

class PMultiplicativeSemigroup (a :: S -> Type) where

The multiplication operation.

Laws

x #* (y #* z) = (x #* y) #* z (associativity of #*)

If you define a custom ppowPositive, ensure the following also hold:

ppowPositive x pone = x

(ppowPositive x n) #* (ppowPositive x m) = ppowPositive x (n #+ m)

ppowPositive (ppowPositive x n) m = ppowPositive x (n #* m)

The default implementation ensures these laws are satisfied.

Note

Unlike PAdditiveSemigroup, the multiplication operation doesn't need to be commutative. Currently, all Plutarch-provided instances are, but this need not be true for other instances.

Since: 1.10.0

Minimal complete definition

Nothing

Methods

(#*) :: forall (s :: S). Term s a -> Term s a -> Term s a infix 6

default (#*) :: forall (s :: S). PMultiplicativeSemigroup (PInner a) => Term s a -> Term s a -> Term s a

ppowPositive :: forall (s :: S). Term s a -> Term s PPositive -> Term s a

class PMultiplicativeSemigroup a => PMultiplicativeMonoid (a :: S -> Type) where

The notion of one (multiplicative identity), and exponentiation by - naturals.

Laws

pone #* x = x (pone is the left identity of #*)

x #* pone = x (pone is the right identity of #*)

ppowPositive pone p = pone (pone does not scale up)

If you define ppowNatural, ensure the following as well:

ppowNatural x (ppositiveToNatural # p) = ppowPositive x p

ppowNatural x pzero = pone

Since: 1.10.0

Minimal complete definition

pone

Methods

pone :: forall (s :: S). Term s a

ppowNatural :: forall (s :: S). Term s a -> Term s PNatural -> Term s a

class (PAdditiveGroup a, PMultiplicativeMonoid a) => PRing (a :: S -> Type) where

Laws

pfromInteger 0 = pzero

pfromInteger 1 = pone

pfromInteger (x + y) = pfromInteger x #+ pfromInteger y

pfromInteger (x * y) = pfromInteger x #* pfromInteger y

Additionally, the following 'interaction laws' must hold between the instances of PAdditiveGroup and PMultiplicativeMonoid for a:

x #* (y #+ z) = (x #* y) #+ (x #* z) (#* left-distributes over #+)

(y #+ z) #* x = (y #* x) #+ (z #* x) (#* right-distributes over #+)

Since: 1.10.0

Minimal complete definition

Nothing

Methods

pfromInteger :: forall (s :: S). Integer -> Term s a

default pfromInteger :: forall (s :: S). PRing (PInner a) => Integer -> Term s a

class (PRing a, POrd a) => PIntegralDomain (a :: S -> Type) where

Laws

Pedantry note

Technically, the requirements here are too strong: we demand an ordered ring, which integral domains don't necessarily have to be. However, in our case, our hand is forced by expected semantics: in abstract algebra, both the absolute value and the signum are real numbers (which are always totally ordered) but in our case, both must be elements of the integral domain itself. Thus, in order for the laws to make any sense, we have to ensure a total order on the integral domain. Since all of our integral domains are 'at least as big' as the integers, this doesn't pose a huge problem.

Since: 1.10.0

Minimal complete definition

Nothing

Methods

psignum :: forall (s :: S). Term s (a :--> a)

default psignum :: forall (s :: S). Term s (a :--> a)

pabs :: forall (s :: S). Term s (a :--> a)

default pabs :: forall (s :: S). Term s (a :--> a)

Functions

positiveToInteger :: Positive -> Integer

Since: 1.10.0

toPositiveAbs :: Integer -> Positive

Converts negative Integers into their absolute values, positive Integers into their Positive equivalents. Errors on 0.

Since: 1.10.0

ptryPositive :: forall (s :: S). Term s (PInteger :--> PPositive)

Partial version of ppositive. Errors if argument is not positive.

Since: 1.10.0

ppositive :: Term s (PInteger :--> PMaybe PPositive)

Build a PPositive from a PInteger. Yields PNothing if argument is not positive.

ptryNatural :: forall (s :: S). Term s (PInteger :--> PNatural)

Partial version of pnatural. Errors if argument is negative.

Since: 1.10.0

pnatural :: forall (s :: S). Term s (PInteger :--> PMaybe PNatural)

Build a PNatural from a PInteger. Yields PNothing if given a negative value.

Since: 1.10.0

ppositiveToNatural :: forall (s :: S). Term s (PPositive :--> PNatural)

'Relax' a PPositive to PNatural. This uses punsafeCoerce underneath, but because any positive is also a natural, is safe.

Since: 1.10.0

pnaturalToPositiveCPS :: forall (a :: S -> Type) (s :: S). Term s a -> (Term s PPositive -> Term s a) -> Term s PNatural -> Term s a

Specialized form of pmaybe for PNatural. Given a default, and a way to turn a PPositive into an answer, produce the default when given pzero, and apply the function otherwise.

Since: 1.10.0

pbySquaringDefault :: forall (a :: S -> Type) (s :: S). (forall (s' :: S). Term s' a -> Term s' a -> Term s' a) -> Term s a -> Term s PPositive -> Term s a

A default implementation of exponentiation-by-squaring with a strictly-positive exponent.

Important note

This implementation assumes that the operation argument is associative.

Since: 1.10.0

pdiv :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

Since: 1.10.0

pmod :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

Since: 1.10.0

pquot :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

Since: 1.10.0

prem :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

Since: 1.10.0

Orphan instances

PIntegralDomain a => Num (Term s a)

Instance details

Methods

(+) :: Term s a -> Term s a -> Term s a

(-) :: Term s a -> Term s a -> Term s a

(*) :: Term s a -> Term s a -> Term s a

negate :: Term s a -> Term s a

abs :: Term s a -> Term s a

signum :: Term s a -> Term s a

fromInteger :: Integer -> Term s a