Plutarch.Internal.PLam

Documentation

plam :: forall c. PLamN a b s => (Term s c -> a) -> Term s (c :--> b)

(#$) :: HasCallStack => Term s (a :--> b) -> Term s a -> Term s b infixr 0

Low precedence infixr synonym of papp, to be used like $, in combination with #. e.g.:

>>> f # x #$ g # y # z
f x (g y z)

(#) :: HasCallStack => Term s (a :--> b) -> Term s a -> Term s b infixl 8

High precedence infixl synonym of papp, to be used like function juxtaposition. e.g.:

>>> f # x # y
f x y

pinl :: Term s a -> (Term s a -> Term s b) -> Term s b