Plutarch.Internal.Semigroup

Type classes

class PSemigroup (a :: S -> Type) where

Laws

The only mandatory law is that #<> must be associative:

x #<> (y #<> z) = (x #<> y) #<> z

If you define pstimes, ensure the following also hold:

pstimes pone x = x

(pstimes p1 x) #<> (pstimes p2 x) = pstimes (p1 #+ p2) x

pstimes p1 (pstimes p2 x) = pstimes (p1 #* p2) x

The default implementation automatically ensures these laws hold.

@since WIP

Minimal complete definition

Nothing

Methods

(#<>) :: forall (s :: S). Term s a -> Term s a -> Term s a infixr 6

default (#<>) :: forall (s :: S). PSemigroup (PInner a) => Term s a -> Term s a -> Term s a

pstimes :: forall (s :: S). Term s PPositive -> Term s a -> Term s a

class PSemigroup a => PMonoid (a :: S -> Type) where

Laws

pmempty #<> x = x #<> pmempty = x

pstimes n pmempty = pmempty

If you define pmtimes, ensure the following as well:

pmtimes (ppositiveToNatural # p) x = pstimes p x

pmtimes pzero x = pmempty

The default implementation of pmtimes ensures these laws hold.

@since WIP

Minimal complete definition

Nothing

Methods

pmempty :: forall (s :: S). Term s a

default pmempty :: forall (s :: S). PMonoid (PInner a) => Term s a

pmtimes :: forall (s :: S). Term s PNatural -> Term s a -> Term s a

Helper newtypes

newtype PAnd (a :: S -> Type) (s :: S)

Wrapper for types which have logical AND semantics somehow.

@since WIP

Constructors

PAnd (Term s a)

newtype POr (a :: S -> Type) (s :: S)

Wrapper for types which have logical OR semantics somehow.

@since WIP

Constructors

POr (Term s a)

newtype PXor (a :: S -> Type) (s :: S)

Wrapper for types which have logical XOR semantics somehow.

@since WIP

Constructors

PXor (Term s a)