Plutarch.Internal.Show

Documentation

class PShow t where

Minimal complete definition

Nothing

Methods

pshow' :: Bool -> Term s t -> Term s PString

Return the string representation of a Plutarch value

If the wrap argument is True, optionally wrap the output in `(..)` if it represents multiple parameters.

default pshow' :: (PGeneric t, PlutusType t, All2 PShow (PCode t)) => Bool -> Term s t -> Term s PString

pshow :: PShow a => Term s a -> Term s PString

Return the string representation of a Plutarch value

pshowAndErr :: Term s a -> Term s b

Causes an error where the input is shown in the message. Works for all types.

pshowList :: forall list a s. (PShow a, PIsListLike list a) => Term s (list a :--> PString)