Plutarch.Internal.Trace

Documentation

ptraceInfo :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Trace the given message at the info level before evaluating the given argument.

Since: 1.6.0

ptraceDebug :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Trace the given message at the debug level before evaluating the given argument.

Since: 1.6.0

ptrace' :: Term s (PString :--> (a :--> a))

ptrace :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Deprecated: Use ptraceInfo

Backward compatibility synonym for ptraceInfo.

Since: 1.6.0