Plutarch.List

Description

Scott-encoded lists and ListLike typeclass

Documentation

data PList (a :: S -> Type) (s :: S)

SOP-encoded list.

@since WIP

Constructors

PSCons (Term s a) (Term s (PList a))

PSNil

ptryUncons :: PIsListLike list a => Term s (list a :--> PPair a (list a))

Extract head and tail of the list, throws error if list is empty.

puncons :: PIsListLike list a => Term s (list a :--> PMaybe (PPair a (list a)))

Extract head and tail of the list, if list is not empty.

pzip :: (PListLike list, PElemConstraint list a, PElemConstraint list b, PElemConstraint list (PPair a b)) => Term s (list a :--> (list b :--> list (PPair a b)))

O(min(n, m)) . Zip two lists together, creating pairs of the elements.

If the lists are of differing lengths, cut to the shortest.

pfind :: PIsListLike l a => Term s ((a :--> PBool) :--> (l a :--> PMaybe a))

O(n) . like haskell level find but on plutarch level

preverse :: forall (l :: (S -> Type) -> S -> Type) (a :: S -> Type) (s :: S). PIsListLike l a => Term s (l a :--> l a)

O(n) . Reverse a list-like structure.

@since WIP

pcheckSorted :: forall (l :: (S -> Type) -> S -> Type) (a :: S -> Type) (s :: S). (PIsListLike l a, POrd a) => Term s (l a :--> PBool)

O(n) . Checks if a list-list structure is sorted.

@since WIP

pelem :: (PIsListLike list a, PEq a) => Term s (a :--> (list a :--> PBool))

O(n) . Check if element is in the list

(#!!) :: PIsListLike l a => Term s (l a) -> Term s PInteger -> Term s a

O(n) . Like Haskell level (!!) but on the plutarch level

pelemAt :: PIsListLike l a => Term s (PInteger :--> (l a :--> a))

O(n) . Like Haskell level (!!) but on the Plutarch level, not infix and with arguments reversed, errors if the specified index is greater than or equal to the lists length

pelemAt' :: PIsListLike l a => Term s (PInteger :--> (l a :--> a))

O(n) . like pelemAt but doesn't fail on negative indexes

plistEquals :: (PIsListLike list a, PEq a) => Term s (list a :--> (list a :--> PBool))

O(min(n, m)) . Check if two lists are equal.