Plutarch.Maybe

Type

data PMaybe (a :: S -> Type) (s :: S)

Plutus Maybe type, with Scott-encoded repr

Constructors

PJust (Term s a)

PNothing

Functions

Introduction

pjust :: forall (a :: S -> Type) (s :: S). Term s (a :--> PMaybe a)

Construct a PJust value.

@since WIP

pnothing :: forall (a :: S -> Type) (s :: S). Term s (PMaybe a)

Construct a PNothing value.

@since WIP

Predicates

pisJust :: forall (a :: S -> Type) (s :: S). Term s (PMaybe a :--> PBool)

Yields true if the given PMaybe value is of form PJust _.

@since WIP

Eliminators

pfromJust :: forall (a :: S -> Type) (s :: S). Term s (PMaybe a :--> a)

Extracts the element out of a PJust and throws an error if its argument is PNothing.

ptraceIfNothing

Arguments

:: forall (a :: S -> Type) (s :: S). Term s PString
The custom error message.

-> Term s (PMaybe a)

-> Term s a

Extracts the element out of a PJust and throws a custom error if it's given a PNothing.

@since WIP

pfromMaybe :: forall (a :: S -> Type) (s :: S). Term s (a :--> (PMaybe a :--> a))

Extract a PMaybe by providing a default value in case of PJust.

@since WIP

pmaybe :: forall (b :: S -> Type) (a :: S -> Type) (s :: S). Term s (b :--> ((a :--> b) :--> (PMaybe a :--> b)))

Given a default value, a function and a PMaybe value, yields the default value if the PMaybe value is PNothing and applies the function to the value stored in the PJust otherwise.

@since WIP

passertPJust :: forall (a :: S -> Type) (s :: S). Term s (PString :--> (PMaybe a :--> a))

Extract the value stored in a PMaybe container. If there's no value, throw an error with the given message.

@since WIP

pmapMaybe :: Term s ((a :--> b) :--> (PMaybe a :--> PMaybe b))

Map underlying value if PMaybe is PJust, do nothing if it is PNothing

@since WIP