Plutarch.Prelude

Documentation

newtype PDataNewtype (a :: S -> Type) (s :: S)

Since: 1.7.0

Constructors

PDataNewtype (Term s (PAsData a))

Boolean

data PBool (s :: S)

Builtin Plutus boolean.

@since WIP

Constructors

PTrue

PFalse

pand' :: forall (s :: S). Term s (PBool :--> (PBool :--> PBool))

As pand, but strict.

@since WIP

pcond :: forall (a :: S -> Type) (s :: S). [(Term s PBool, Term s a)] -> Term s a -> Term s a

Essentially multi-way pif. More precisely, given a list of condition-action pairs, and an 'action of last resort', construct a left-to-right 'chain' of pifs, using the conditions to determine which action gets taken. The 'action of last resort' finishes the 'chain'. For example:

pcond [(cond1, act1), (cond2, act2)] act3

does the same thing as

pif cond1 act1 (pif cond2 act2 act3)

@since WIP

pif :: forall (a :: S -> Type) (s :: S). Term s PBool -> Term s a -> Term s a -> Term s a

Lazy if-then-else.

@since WIP

pif' :: forall (a :: S -> Type) (s :: S). Term s (PBool :--> (a :--> (a :--> a)))

Strict if-then-else. Emits slightly less code than the lazy version.

@since WIP

pnot :: forall (s :: S). Term s (PBool :--> PBool)

Boolean negation.

@since WIP

por' :: Term s (PBool :--> (PBool :--> PBool))

As por, but strict.

@since WIP

(#&&) :: forall (s :: S). Term s PBool -> Term s PBool -> Term s PBool infixr 3

Lazy AND for terms.

@since WIP

(#||) :: forall (s :: S). Term s PBool -> Term s PBool -> Term s PBool infixr 2

Lazy OR for terms.

@since WIP

Bytestring

data PByte (s :: S)

A Plutarch-level representation of bytes.

Note =

This type is intentionally quite restrictive, as it's not really meant to be computed with. Instead, it ensures certain operations' type safety while also allowing more sensible signatures. If you want to do anything with PBytes, we recommend converting them to PIntegers first.

@since WIP

data PByteString s

Plutus BuiltinByteString

data PLogicOpSemantics (s :: S)

Type designating whether logical operations should use padding or truncation semantics. See CIP-122 for more details on this.

@since WIP

pandBS :: forall (s :: S). Term s (PLogicOpSemantics :--> (PByteString :--> (PByteString :--> PByteString)))

Perform the logical AND of two PByteStrings, as per CIP-122. The PLogicOpSemantics argument specifies what should be done if the lengths of the two PByteString arguments do not match.

@since WIP

pbyteToInteger :: Term s (PByte :--> PInteger)

Convert a PByte into its corresponding PInteger.

@since WIP

pcomplementBS :: forall (s :: S). Term s (PByteString :--> PByteString)

Perform the logical complement of a PByteString, as per CIP-122.

@since WIP

pconsBS :: Term s (PByte :--> (PByteString :--> PByteString))

Prepend a PByte to a 'PByteString.

@since WIP

phexByteStr :: HasCallStack => String -> Term s PByteString

Interpret a hex string as a PByteString.

pindexBS :: Term s (PByteString :--> (PInteger :--> PByte))

Given a valid index into a PByteString, returns the PByte at that index. Will crash if given an out-of-bounds index.

@since WIP

pintegerToByte :: Term s (PInteger :--> PByte)

Try to convert a PInteger into its corresponding PByte. This operation unchecked: use with care.

@since WIP

plengthBS :: Term s (PByteString :--> PInteger)

Find the length of a PByteString.

porBS :: forall (s :: S). Term s (PLogicOpSemantics :--> (PByteString :--> (PByteString :--> PByteString)))

Perform the logical OR of two PByteStrings, as per CIP-122. The PLogicOpSemantics argument specifies what should be done if the lengths of the two PByteString arguments do not match.

@since WIP

ppadding :: forall (s :: S). Term s PLogicOpSemantics

Indicates that padding semantics should be used.

@since WIP

preplicateBS :: forall (s :: S). Term s (PInteger :--> (PByte :--> PByteString))

Given a desired length and a PByte, construct a PByteString of the specified length (0 if negative) consisting entirely of that PByte.

@since WIP

psliceBS :: Term s (PInteger :--> (PInteger :--> (PByteString :--> PByteString)))

Slice a PByteString with given start index and slice length.

>>> (pslice # 2 # 3 phexByteStr "4102afde5b2a") #== phexByteStr "afde5b"

ptruncation :: forall (s :: S). Term s PLogicOpSemantics

Indicates that truncation semantics should be used.

@since WIP

pxorBS :: forall (s :: S). Term s (PLogicOpSemantics :--> (PByteString :--> (PByteString :--> PByteString)))

Perform the logical XOR of two PByteStrings, as per CIP-122. The PLogicOpSemantics argument specifies what should be done if the lengths of the two PByteString arguments do not match.

@since WIP

Cryptographic primitives

psha2_256 :: Term s (PByteString :--> PByteString)

Hash a PByteString using SHA-256.

psha3_256 :: Term s (PByteString :--> PByteString)

Hash a PByteString using SHA3-256.

pverifySignature :: Term s (PByteString :--> (PByteString :--> (PByteString :--> PBool)))

Deprecated: use one of the Ed25519, Schnorr- or ECDSA Secp256k1 signature verification functions

Verify the signature against the public key and message.

Data encoding

newtype PAsData (a :: S -> Type) (s :: S)

Constructors

PAsData (Term s a)

data PBuiltinList (a :: S -> Type) (s :: S)

Plutus BuiltinList

Constructors

PCons (Term s a) (Term s (PBuiltinList a))

PNil

newtype PBuiltinPair (a :: S -> Type) (b :: S -> Type) (s :: S)

Constructors

PBuiltinPair (Term s (PBuiltinPair a b))

newtype PData (s :: S)

Constructors

PData (Term s PData)

pasByteStr :: Term s (PData :--> PByteString)

pasConstr :: Term s (PData :--> PBuiltinPair PInteger (PBuiltinList PData))

pasInt :: Term s (PData :--> PInteger)

pasList :: Term s (PData :--> PBuiltinList PData)

pasMap :: Term s (PData :--> PBuiltinList (PBuiltinPair PData PData))

pchooseData :: Term s (PData :--> (a :--> (a :--> (a :--> (a :--> (a :--> a))))))

pchooseListBuiltin :: Term s (PBuiltinList a :--> (b :--> (b :--> b)))

pconsBuiltin :: Term s (a :--> (PBuiltinList a :--> PBuiltinList a))

pconstrBuiltin :: Term s (PInteger :--> (PBuiltinList PData :--> PAsData (PBuiltinPair PInteger (PBuiltinList PData))))

pfstBuiltin :: Term s (PBuiltinPair a b :--> a)

pheadBuiltin :: Term s (PBuiltinList a :--> a)

plistData :: Term s (PBuiltinList PData :--> PData)

pnullBuiltin :: Term s (PBuiltinList a :--> PBool)

ppairDataBuiltin :: Term s (PAsData a :--> (PAsData b :--> PBuiltinPair (PAsData a) (PAsData b)))

Construct a builtin pair of PData elements.

Uses PAsData to preserve more information about the underlying PData.

pserialiseData :: Term s (PData :--> PByteString)

Serialise any builtin data to its cbor represented by a builtin bytestring

psndBuiltin :: Term s (PBuiltinPair a b :--> b)

ptailBuiltin :: Term s (PBuiltinList a :--> PBuiltinList a)

Integer

data PInteger s

A builtin Plutus integer.

@since WIP

Opaque

newtype POpaque s

An Arbitrary Term with an unknown type

Constructors

POpaque (Term s POpaque)

popaque :: Term s a -> Term s POpaque

Erase the type of a Term

String

data PString s

Plutus BuiltinString values

pdecodeUtf8 :: Term s (PByteString :--> PString)

Decode a PByteString using UTF-8.

pencodeUtf8 :: Term s (PString :--> PByteString)

Encode a PString using UTF-8.

Unit

data PUnit (s :: S)

Constructors

PUnit

DataRepr

class PDataFields (a :: S -> Type)

Class allowing letFields to work for a PType, usually via PIsDataRepr, but is derived for some other types for convenience.

data PDataRecord (as :: [PLabeledType]) (s :: S)

A "record" of `exists a. PAsData a`. The underlying representation is `PBuiltinList PData`.

data PDataSum defs s

A sum of PDataRecords. The underlying representation is the Constr constructor, where the integer is the index of the variant and the list is the record.

data PLabeledType

Constructors

Symbol := (S -> Type)

data PlutusTypeData

pdcons :: forall label a l s. Term s (PAsData a :--> (PDataRecord l :--> PDataRecord ((label ':= a) ': l)))

Cons a field to a data record.

You can specify the label to associate with the field using type applications-

foo :: Term s (PDataRecord '[ "fooField" ':= PByteString ])
foo = pdcons @"fooField" # pdata (phexByteStr "ab") # pdnil

pdnil :: Term s (PDataRecord '[])

An empty PDataRecord.

pfield :: forall name b p s a as n. (PDataFields p, as ~ PFields p, n ~ PLabelIndex name as, KnownNat n, a ~ PUnLabel (IndexList n as), PFromDataable a b) => Term s (p :--> b)

Get a single field from a Term.

NB*: If you access more than one field from the same value you should use pletFields instead, which will generate the bindings more efficiently.

pletFields :: forall fs a s b ps bs. (PDataFields a, ps ~ PFields a, bs ~ Bindings ps fs, BindFields ps bs) => Term s a -> (HRecOf a fs s -> Term s b) -> Term s b

Bind a HRec of named fields containing all the specified fields.

Either

data PEither (a :: S -> Type) (b :: S -> Type) (s :: S)

SOP-encoded Either.

@since WIP

Constructors

PLeft (Term s a)

PRight (Term s b)

Enumerable and Countable

class POrd a => PCountable (a :: S -> Type) where

A notion of 'next' value. More formally, instances of this type class are discrete linear orders with no maximal element.

Laws

x /= psuccessor x

y < x = psuccessor y <= x

x < psuccessor y = x <= y

If you define psuccessorN, you must also ensure the following hold; the default implementation ensures this.

psuccessorN 1 = psuccessor

psuccessorN n . psuccessorN m = psuccessorN (n + m)

Law 1 ensures no value is its own successor. Laws 2 and 3 ensure that there are no 'gaps': every value is 'reachable' from any lower value by a finite number of applications of successor.

@since WIP

Minimal complete definition

psuccessor

Methods

psuccessor :: forall (s :: S). Term s (a :--> a)

@since WIP

psuccessorN :: forall (s :: S). Term s (PPositive :--> (a :--> a))

The default implementation of this function is inefficient: if at all possible, give instances an optimized version that doesn't require recursion.

@since WIP

class PCountable a => PEnumerable (a :: S -> Type) where

Similar to PCountable, but has the ability to get a 'previous' value as well. More formally, instances of this type class are discrete linear orders with no maximal or minimal element.

Laws

ppredecessor . psuccessor = psuccessor . ppredecessor = id

If you define ppredecessorN, you must also ensure the following hold; the default implementation ensures this.

ppredecessorN 1 = ppredecessor

ppredecessorN n . ppredecessorN m = ppredecessorN (n + m)

From Law 1, we obtain the following theorem:

x /= predecessor x

@since WIP

Minimal complete definition

ppredecessor

Methods

ppredecessor :: forall (s :: S). Term s (a :--> a)

@since WIP

ppredecessorN :: forall (s :: S). Term s (PPositive :--> (a :--> a))

The default implementation of this function is inefficient: if at all possible, give instances an optimized version that doesn't require recursion.

@since WIP

Eq and Ord

class PEq t where

Minimal complete definition

Nothing

Methods

(#==) :: Term s t -> Term s t -> Term s PBool infix 4

default (#==) :: (PGeneric t, PlutusType t, All2 PEq (PCode t)) => Term s t -> Term s t -> Term s PBool

class PEq t => POrd t where

Total ordering relation.

Laws

#<= must form a total order. More precisely:

x #<= x = pcon PTrue (reflexivity)

(y #< x) #|| (z #< y) #|| (x #<= z) = pcon PTrue (transitivity)

(x #<= y) #|| (y #<= x) = pcon PTrue (totality)

Furthermore, #< must be an equivalent strict total order to #<=:

x #< x = pcon PFalse (irreflexivity)

(y #<= x) #|| (z #<= y) #|| (x #< z) = pcon PTrue (transitivity)

(x #< y) #|| (y #< x) #|| (x #== z) = pcon PTrue (trichotomy)

x #<= y = (x #< y) #|| (x #== y) (strict equivalence)

If you define pmax or pmin, ensure the following also hold:

pmax # x # y = pmax # y # x (commutativity, also for @pmin)

pmax # x #$ pmax y z = pmax # (pmax # x # y) # z (associativity, also for @pmin)

pmax # x #$ pmin # y # z = pmin # (pmax # x # y) # (pmax # x # z) (pmax distributes over pmin, also equivalent for pmin)

pmin x y = pif' (x #<= y) x y

pmax x y = pif' (x #<= y) y x

Laws 8-12 hold if you use the defaults provided by this type class.

@since WIP

Minimal complete definition

Nothing

Methods

(#<=) :: Term s t -> Term s t -> Term s PBool infix 4

@since WIP

default (#<=) :: POrd (PInner t) => Term s t -> Term s t -> Term s PBool

(#<) :: Term s t -> Term s t -> Term s PBool infix 4

@since WIP

default (#<) :: POrd (PInner t) => Term s t -> Term s t -> Term s PBool

pmax :: forall (s :: S). Term s t -> Term s t -> Term s t

@since WIP

pmin :: forall (s :: S). Term s t -> Term s t -> Term s t

@since WIP

(#>) :: forall (a :: S -> Type) (s :: S). POrd a => Term s a -> Term s a -> Term s PBool infix 4

@since WIP

(#>=) :: forall (a :: S -> Type) (s :: S). POrd a => Term s a -> Term s a -> Term s PBool infix 4

@since WIP

Fixed point

pfix :: Term s (((a :--> b) :--> (a :--> b)) :--> (a :--> b))

Fixpoint recursion. Used to encode recursive functions.

Example:

iterateN' ::
 Term s (PInteger :--> (a :--> a) :--> a :--> a) ->
 Term s PInteger ->
 Term s (a :--> a) ->
 Term s a
iterateN' self n f x =
  pif (n #== 0) x (self # n - 1 #$ f x)

iterateN :: Term s (PInteger :--> (a :--> a) :--> a :--> a)
iterateN = pfix #$ plam iterateN'

Further examples can be found in examples/Recursion.hs

IsData

class PIsData a where

Laws: - If PSubtype PData a, then pdataImpl a must be pupcast. - pdataImpl . pupcast . pfromDataImpl ≡ id - pfromDataImpl . punsafeDowncast . pdataImpl ≡ id

Minimal complete definition

Nothing

Methods

pfromDataImpl :: Term s (PAsData a) -> Term s a

default pfromDataImpl :: PIsData (PInner a) => Term s (PAsData a) -> Term s a

pdataImpl :: Term s a -> Term s PData

default pdataImpl :: PIsData (PInner a) => Term s a -> Term s PData

pdata :: PIsData a => Term s a -> Term s (PAsData a)

pforgetData :: forall s a. Term s (PAsData a) -> Term s PData

pfromData :: PIsData a => Term s (PAsData a) -> Term s a

Lifting and lowering

class PlutusType a => PLiftable (a :: S -> Type) where

Indicates that the given Plutarch type has an equivalent in Haskell (and Plutus by extension), and we have the ability to move between them.

Important note

Calling methods of PLiftable directly should rarely, if ever, be a thing you do, unless defining your own instances without via-deriving helpers (below). Prefer using pconstant and plift, as these handle some of the oddities required without you having to think about them.

You should rarely, if ever, need to define PLiftable instances by hand. Whenever possible, prefer using DeriveBuiltinPLiftable, DeriveDataPLiftable, and DeriveNewtypePLiftable as they have fewer complexities and caveats. See their documentation for when to use them.

If you do want to define the methods yourself, there's a few key factors to keep in mind:

You still shouldn't write every method by hand, there are helpers fromPlutarch* and toPlutarch* to cover common cases like types in Plutus universe or Scott encoding

If defining toPlutarchRepr and fromPlutarchRepr you will need to define an associated PlutusRepr type, this is a Hasekll level type that is included in the Plutus default universe.

If defining toPlutarch and fromPlutarch for Scott encoded type you need to set PlutusRepr PMyType = PLiftedClosed PMyType

When choosing a type for AsHaskell, any value of that type must be representable in Plutarch. If you have internal invariants to maintain on the Haskell side, make sure you do so with great care.

Laws

fromPlutarchRepr . toPlutarchRepr = Just

fmap toPlutarchRepr . fromPlutarchRepr = Just

fromPlutarch . toPlutarch = Right

fmap toPlutarch . fromPlutarch = Right

Any derivations via DeriveBuiltinPLiftable, DeriveDataPLiftable, and DeriveNewtypePLiftable automatically follow these laws.

@since WIP

Associated Types

type AsHaskell a :: Type

type PlutusRepr a :: Type

Methods

toPlutarchRepr :: AsHaskell a -> PlutusRepr a

toPlutarch :: AsHaskell a -> PLifted s a

fromPlutarchRepr :: PlutusRepr a -> Maybe (AsHaskell a)

fromPlutarch :: (forall s. PLifted s a) -> Either LiftError (AsHaskell a)

data DeriveDataPLiftable (a :: S -> Type) (h :: Type) (s :: S)

via-deriving helper, indicating that a has a Haskell-level equivalent h by way of its Data encoding, rather than by h being directly part of the Plutus default universe.

@since WIP

data DeriveNewtypePLiftable (wrapper :: S -> Type) (inner :: S -> Type) (h :: Type) (s :: S)

via-deriving helper, indicating that wrapper has a Haskell-level equivalent h by way of encoding of inner. It requires that AsHaskell inner has the same Haskell representation as h

@since WIP

newtype PLifted (s :: S) (a :: S -> Type)

Similar to Identity, but at the level of Plutarch. Only needed when writing manual instances of PLiftable, or if you want to use toPlutarch and fromPlutarch directly.

This is used for coercing Plutarch terms in Haskell level with `coerce :: PLifted s a -> PLifted s b` for via-deriving helpers

@since WIP

Constructors

PLifted (Term s POpaque)

fromPlutarchUni :: forall (a :: S -> Type). (PLiftable a, DefaultUni `Includes` PlutusRepr a) => (forall s. PLifted s a) -> Either LiftError (AsHaskell a)

Valid definition for fromPlutarch if PlutusRepr is in Plutus universe

@since WIP

toPlutarchUni :: forall (a :: S -> Type) (s :: S). (PLiftable a, DefaultUni `Includes` PlutusRepr a) => AsHaskell a -> PLifted s a

Valid definition for toPlutarch if PlutusRepr is in Plutus universe

@since WIP

pconstant :: forall (a :: S -> Type) (s :: S). PLiftable a => AsHaskell a -> Term s a

Given a Haskell-level representation of a Plutarch term, transform it into its equivalent term.

@since WIP

plift :: forall (a :: S -> Type). PLiftable a => (forall (s :: S). Term s a) -> AsHaskell a

Given a closed Plutarch term, compile and evaluate it, then produce the corresponding Haskell value. If compilation or evaluation fails somehow, this will call error: if you need to 'trap' these outcomes and handle them differently somehow, use fromPlutarch.

@since WIP

Lists

type family PElemConstraint list (a :: S -> Type) :: Constraint

type PIsListLike list a = (PListLike list, PElemConstraint list a)

'PIsListLike list a' constraints list be a PListLike with valid element type, a.

class PListLike (list :: (S -> Type) -> S -> Type)

Plutarch types that behave like lists.

Minimal complete definition

pelimList, pcons, pnil

data PList (a :: S -> Type) (s :: S)

SOP-encoded list.

@since WIP

Constructors

PSCons (Term s a) (Term s (PList a))

PSNil

pelem :: (PIsListLike list a, PEq a) => Term s (a :--> (list a :--> PBool))

O(n) . Check if element is in the list

pelemAt :: PIsListLike l a => Term s (PInteger :--> (l a :--> a))

O(n) . Like Haskell level (!!) but on the Plutarch level, not infix and with arguments reversed, errors if the specified index is greater than or equal to the lists length

pfind :: PIsListLike l a => Term s ((a :--> PBool) :--> (l a :--> PMaybe a))

O(n) . like haskell level find but on plutarch level

plistEquals :: (PIsListLike list a, PEq a) => Term s (list a :--> (list a :--> PBool))

O(min(n, m)) . Check if two lists are equal.

puncons :: PIsListLike list a => Term s (list a :--> PMaybe (PPair a (list a)))

Extract head and tail of the list, if list is not empty.

pzip :: (PListLike list, PElemConstraint list a, PElemConstraint list b, PElemConstraint list (PPair a b)) => Term s (list a :--> (list b :--> list (PPair a b)))

O(min(n, m)) . Zip two lists together, creating pairs of the elements.

If the lists are of differing lengths, cut to the shortest.

(#!!) :: PIsListLike l a => Term s (l a) -> Term s PInteger -> Term s a

O(n) . Like Haskell level (!!) but on the plutarch level

pall :: PIsListLike list a => Term s ((a :--> PBool) :--> (list a :--> PBool))

O(n) . Check that predicate holds for all elements in a list.

pany :: PIsListLike list a => Term s ((a :--> PBool) :--> (list a :--> PBool))

O(n) . Check that predicate holds for any element in a list.

pconcat :: PIsListLike list a => Term s (list a :--> (list a :--> list a))

O(n) . Concatenate two lists

Example: > pconcat # psingleton x # psingleton y == plistLiteral [x, y]

pconcat exhibits identities with empty lists such that > forall x. pconcat # pnil # x == x > forall x. pconcat # x # pnil == x

pcons :: (PListLike list, PElemConstraint list a) => Term s (a :--> (list a :--> list a))

Cons an element onto an existing list.

pdrop :: PIsListLike list a => Natural -> Term s (list a) -> Term s (list a)

Drop the first n fields of a List.

The term will be statically generated as repeated applications of ptail, which will be more efficient in many circumstances.

pelimList :: (PListLike list, PElemConstraint list a) => (Term s a -> Term s (list a) -> Term s r) -> Term s r -> Term s (list a) -> Term s r

Canonical eliminator for list-likes.

pfilter :: PIsListLike list a => Term s ((a :--> PBool) :--> (list a :--> list a))

O(n) . Filter elements from a list that don't match the predicate.

pfoldl :: PIsListLike list a => Term s ((b :--> (a :--> b)) :--> (b :--> (list a :--> b)))

O(n) . Fold on a list left-associatively.

pfoldr :: PIsListLike list a => Term s ((a :--> (b :--> b)) :--> (b :--> (list a :--> b)))

O(n) . Fold on a list right-associatively.

pfoldrLazy :: PIsListLike list a => Term s ((a :--> (PDelayed b :--> b)) :--> (b :--> (list a :--> b)))

O(n) . Fold on a list right-associatively, with opportunity for short circuting.

May short circuit when given reducer function is lazy in its second argument.

phead :: (PListLike list, PElemConstraint list a) => Term s (list a :--> a)

Return the first element of a list. Partial, throws an error upon encountering an empty list.

plength :: PIsListLike list a => Term s (list a :--> PInteger)

O(n) . Count the number of elements in the list

pmap :: (PListLike list, PElemConstraint list a, PElemConstraint list b) => Term s ((a :--> b) :--> (list a :--> list b))

O(n) . Map a function over a list of elements

pnil :: (PListLike list, PElemConstraint list a) => Term s (list a)

The empty list

pnull :: (PListLike list, PElemConstraint list a) => Term s (list a :--> PBool)

O(1) . Check if a list is empty

precList :: PIsListLike list a => (Term s (list a :--> r) -> Term s a -> Term s (list a) -> Term s r) -> (Term s (list a :--> r) -> Term s r) -> Term s (list a :--> r)

Like pelimList, but with a fixpoint recursion hatch.

psingleton :: PIsListLike list a => Term s (a :--> list a)

O(1) . Create a singleton list from an element

ptail :: (PListLike list, PElemConstraint list a) => Term s (list a :--> list a)

Take the tail of a list, meaning drop its head. Partial, throws an error upon encountering an empty list.

ptryIndex :: PIsListLike list a => Natural -> Term s (list a) -> Term s a

Index a BuiltinList, throwing an error if the index is out of bounds.

pzipWith :: (PListLike list, PElemConstraint list a, PElemConstraint list b, PElemConstraint list c) => Term s ((a :--> (b :--> c)) :--> (list a :--> (list b :--> list c)))

O(min(n, m)) . Zip two lists together with a passed function.

If the lists are of differing lengths, cut to the shortest.

pzipWith' :: (PListLike list, PElemConstraint list a, PElemConstraint list b, PElemConstraint list c) => (Term s a -> Term s b -> Term s c) -> Term s (list a :--> (list b :--> list c))

Like pzipWith but with Haskell-level merge function.

PlutusType

class (PInner a ~ DerivedPInner (DPTStrat a) a, PlutusTypeStrat (DPTStrat a), PlutusTypeStratConstraint (DPTStrat a) a, PlutusType a) => DerivePlutusType (a :: PType)

Associated Types

type DPTStrat a :: Type

type DPTStrat a = TypeError ('Text "Please specify a strategy for deriving PlutusType for type " ':<>: 'ShowType a)

type PCon = PlutusType

Deprecated: Use PlutusType

type PMatch = PlutusType

Deprecated: Use PlutusType

class PlutusType (a :: PType)

Associated Types

type PInner a :: PType

type PInner a = DerivedPInner (DPTStrat a) a

pcon :: PlutusType a => a s -> Term s a

Construct a Plutarch Term via a Haskell datatype

pmatch :: PlutusType a => Term s a -> (a s -> Term s b) -> Term s b

Pattern match over Plutarch Terms via a Haskell datatype

data PlutusTypeNewtype

data PlutusTypeScott

Numeric

data Positive

@since WIP

data PPositive (s :: S)

@since WIP

data PNatural (s :: S)

@since WIP

class PAdditiveSemigroup (a :: S -> Type) where

The addition operation, and the notion of scaling by a positive.

Laws

x #+ y = y #+ x (commutativity of #+)

x #+ (y #+ z) = (x #+ y) #+ z (associativity of #+)

If you define a custom pscalePositive, ensure the following also hold:

pscalePositive # x # pone = x

(pscalePositive # x # n) #+ (pscalePositive # x # m) = pscalePositive # x # (n #+ m)

pscalePositive # (pscalePositive # x # n) # m = pscalePositive # x # (n #* m)

The default implementation ensures these laws are satisfied.

@since WIP

Minimal complete definition

Nothing

Methods

(#+) :: forall (s :: S). Term s a -> Term s a -> Term s a infix 6

default (#+) :: forall (s :: S). PAdditiveSemigroup (PInner a) => Term s a -> Term s a -> Term s a

pscalePositive :: forall (s :: S). Term s a -> Term s PPositive -> Term s a

This defaults to exponentiation-by-squaring, which in general is the best we can do.

class PAdditiveSemigroup a => PAdditiveMonoid (a :: S -> Type) where

The notion of zero, as well as a way to scale by naturals.

Laws

pzero #+ x = x (pzero is the identity of #+)

pscalePositive # pzero # n = pzero (pzero does not scale up)

If you define pscaleNatural, ensure the following as well:

pscaleNatural # x #$ ppositiveToNatural # p = pscalePositive # x # p

pscaleNatural # x # pzero = pzero

The default implementation of pscaleNatural ensures these laws are followed.

@since WIP

Minimal complete definition

pzero

Methods

pzero :: forall (s :: S). Term s a

pscaleNatural :: forall (s :: S). Term s a -> Term s PNatural -> Term s a

class PAdditiveMonoid a => PAdditiveGroup (a :: S -> Type) where

The notion of additive inverses, and the subtraction operation.

Laws

If you define pnegate, the following laws must hold:

(pnegate # x) #+ x = pzero (pnegate is an additive inverse)

pnegate #$ pnegate # x = x (pnegate is self-inverting)

If you define #-, the following law must hold:

x #- x = pzero

Additionally, the following 'consistency laws' must hold. Default implementations of both pnegate and #- uphold these.

pnegate # x = pzero #- x

x #- y = x #+ (pnegate # y)

Lastly, if you define a custom pscaleInteger, the following laws must hold:

pscaleInteger # x # pzero = pzero

pscaleInteger # x #$ pnegate y = pnegate #$ pscaleInteger # x # y

@since WIP

Minimal complete definition

pnegate | (#-)

Methods

pnegate :: forall (s :: S). Term s (a :--> a)

(#-) :: forall (s :: S). Term s a -> Term s a -> Term s a infix 6

pscaleInteger :: forall (s :: S). Term s a -> Term s PInteger -> Term s a

class PMultiplicativeSemigroup (a :: S -> Type) where

The multiplication operation.

Laws

x #* (y #* z) = (x #* y) #* z (associativity of #*)

If you define a custom ppowPositive, ensure the following also hold:

ppowPositive # x # pone = x

(ppowPositive # x # n) #* (ppowPositive # x # m) = pscalePositive # x # (n #+ m)

ppowPositive # (ppowPositive # x # n) # m = pscalePositive # x # (n #* m)

The default implementation ensures these laws are satisfied.

Note

Unlike PAdditiveSemigroup, the multiplication operation doesn't need to be commutative. Currently, all Plutarch-provided instances are, but this need not be true for other instances.

@since WIP

Minimal complete definition

Nothing

Methods

(#*) :: forall (s :: S). Term s a -> Term s a -> Term s a infix 6

default (#*) :: forall (s :: S). PMultiplicativeSemigroup (PInner a) => Term s a -> Term s a -> Term s a

ppowPositive :: forall (s :: S). Term s a -> Term s PPositive -> Term s a

class PMultiplicativeSemigroup a => PMultiplicativeMonoid (a :: S -> Type) where

The notion of one (multiplicative identity), and exponentiation by - naturals.

Laws

pone #* x = x (pone is the left identity of #*)

x #* pone = x (pone is the right identity of #*)

ppowPositive # pone # p = pone (pone does not scale up)

If you define ppowNatural, ensure the following as well:

ppowNatural # x #$ ppositiveToNatural # p = ppowPositive # x # p

ppowNatural # x # pzero = pone

@since WIP

Minimal complete definition

pone

Methods

pone :: forall (s :: S). Term s a

ppowNatural :: forall (s :: S). Term s a -> Term s PNatural -> Term s a

class (PAdditiveGroup a, PMultiplicativeMonoid a) => PRing (a :: S -> Type) where

Laws

pfromInteger 0 = pzero

pfromInteger 1 = pone

pfromInteger (x + y) = pfromInteger x #+ pfromInteger y

pfromInteger (x * y) = pfromInteger x #* pfromInteger y

Additionally, the following 'interaction laws' must hold between the instances of PAdditiveGroup and PMultiplicativeMonoid for a:

x #* (y #+ z) = (x #* y) #+ (x #* z) (#* left-distributes over #+)

(y #+ z) #* x = (y #* x) #+ (z #* x) (#* right-distributes over #+)

@since WIP

Minimal complete definition

Nothing

Methods

pfromInteger :: forall (s :: S). Integer -> Term s a

default pfromInteger :: forall (s :: S). PRing (PInner a) => Integer -> Term s a

class (PRing a, POrd a) => PIntegralDomain (a :: S -> Type) where

Laws

Pedantry note

Technically, the requirements here are too strong: we demand an ordered ring, which integral domains don't necessarily have to be. However, in our case, our hand is forced by expected semantics: in abstract algebra, both the absolute value and the signum are real numbers (which are always totally ordered) but in our case, both must be elements of the integral domain itself. Thus, in order for the laws to make any sense, we have to ensure a total order on the integral domain. Since all of our integral domains are 'at least as big' as the integers, this doesn't pose a huge problem.

@since WIP

Minimal complete definition

Nothing

Methods

psignum :: forall (s :: S). Term s (a :--> a)

default psignum :: forall (s :: S). Term s (a :--> a)

pabs :: forall (s :: S). Term s (a :--> a)

default pabs :: forall (s :: S). Term s (a :--> a)

pquot :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

@since WIP

prem :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

@since WIP

pdiv :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

@since WIP

pmod :: forall (s :: S). Term s (PInteger :--> (PInteger :--> PInteger))

@since WIP

ppositive :: Term s (PInteger :--> PMaybe PPositive)

Build a PPositive from a PInteger. Yields PNothing if argument is not positive.

ptryPositive :: forall (s :: S). Term s (PInteger :--> PPositive)

Partial version of ppositive. Errors if argument is not positive.

@since WIP

pnatural :: forall (s :: S). Term s (PInteger :--> PMaybe PNatural)

Build a PNatural from a PInteger. Yields PNothing if given a negative value.

@since WIP

ptryNatural :: forall (s :: S). Term s (PInteger :--> PNatural)

Partial version of pnatural. Errors if argument is negative.

@since WIP

ppositiveToNatural :: forall (s :: S). Term s (PPositive :--> PNatural)

'Relax' a PPositive to PNatural. This uses punsafeCoerce underneath, but because any positive is also a natural, is safe.

@since WIP

Other

pto :: Term s a -> Term s (PInner a)

Safely coerce from a Term to it's PInner representation.

pinl :: Term s a -> (Term s a -> Term s b) -> Term s b

plam :: forall c. (PLamN a b s, HasCallStack) => (Term s c -> a) -> Term s (c :--> b)

newtype PForall (b :: a -> PType) s

Constructors

PForall (forall (x :: a). Term s (b x))

Show

class PShow t

pshow :: PShow a => Term s a -> Term s PString

Return the string representation of a Plutarch value

Term and related functionality

data Term (s :: S) (a :: PType)

data S

type ClosedTerm (a :: PType) = forall (s :: S). Term s a

Closed* terms with no free variables.

data PDelayed (a :: PType) (s :: S)

papp :: Term s (a :--> b) -> Term s a -> Term s b

Lambda Application.

pdelay :: Term s a -> Term s (PDelayed a)

Plutus 'delay', used for laziness.

pforce :: Term s (PDelayed a) -> Term s a

Plutus 'force', used to force evaluation of PDelayed terms.

perror :: Term s a

Plutus 'error'.

When using this explicitly, it should be ensured that the containing term is delayed, avoiding premature evaluation.

phoistAcyclic :: HasCallStack => ClosedTerm a -> Term s a

plet :: Term s a -> (Term s a -> Term s b) -> Term s b

Let bindings.

This is approximately a shorthand for a lambda and application:

plet v f == papp (plam f) v

But sufficiently small terms in WHNF may be inlined for efficiency.

pthrow :: HasCallStack => Text -> Term s a

(#) :: Term s (a :--> b) -> Term s a -> Term s b infixl 8

High precedence infixl synonym of papp, to be used like function juxtaposition. e.g.:

>>> f # x # y
f x y

(#$) :: Term s (a :--> b) -> Term s a -> Term s b infixr 0

Low precedence infixr synonym of papp, to be used like $, in combination with #. e.g.:

>>> f # x #$ g # y # z
f x (g y z)

data ((a :: PType) :--> (b :: PType)) (s :: S) infixr 0

Conversion

type family PSubtype (a :: PType) (b :: PType) :: Constraint where ...

Equations

PSubtype a b = (PSubtype' a b ~ 'PSubtypeRelation, PSubtypeHelper a b (PSubtype' a b))

class PSubtype a b => PTryFrom (a :: PType) (b :: PType) where

PTryFrom a b represents a subtyping relationship between a and b, and a way to go from a to b. Laws: - (punsafeCoerce . fst) $ tcont (ptryFrom x) ≡ pure x

Minimal complete definition

Nothing

Associated Types

type PTryFromExcess a b :: PType

type PTryFromExcess a b = PTryFromExcess a (PInner b)

Methods

ptryFrom' :: forall s r. Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r

default ptryFrom' :: forall s r. (PTryFrom a (PInner b), PTryFromExcess a b ~ PTryFromExcess a (PInner b)) => Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r

ptryFrom :: forall b a s r. PTryFrom a b => Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r

pupcast :: forall a b s. PSubtype a b => Term s b -> Term s a

Maybe

data PMaybe (a :: S -> Type) (s :: S)

@since WIP

Constructors

PJust (Term s a)

PNothing

Pair

data PPair (a :: S -> Type) (b :: S -> Type) (s :: S)

Plutus encoding of Pairs.

Note: This is represented differently than BuiltinPair. It is scott-encoded.

Constructors

PPair (Term s a) (Term s b)

Rational

data PRational s

A Scott-encoded rational number, with a guaranteed positive denominator (and thus, a canonical form).

Note

This is not the Plutarch equivalent of a Plutus Rational; for this, you want PRationalData from plutarch-ledger-api. PRational is designed to optimize for computation: if you want to do any serious work with rational numbers that isn't just passing them around, you want to use (or convert to) PRational.

Constructors

PRational (Term s PInteger) (Term s PPositive)

pdenominator :: Term s (PRational :--> PPositive)

pnumerator :: Term s (PRational :--> PInteger)

pround :: Term s (PRational :--> PInteger)

TermCont

newtype TermCont :: forall (r :: PType). S -> Type -> Type where

Constructors

TermCont

Fields

:: forall r s a. { runTermCont :: (a -> Term s r) -> Term s r

} -> TermCont @r s a

pguardC :: Term s PString -> Term s PBool -> TermCont s ()

Trace a message and raise error if cond is false. Otherwise, continue.

Example ===

onlyAllow42 :: Term s (PInteger :--> PUnit)
onlyAllow42 = plam $ i -> unTermCont $ do
  pguardC "expected 42" $ i #== 42
  pure $ pconstant ()

pguardC' :: Term s a -> Term s PBool -> TermCont @a s ()

Stop computation and return given term if cond is false. Otherwise, continue.

Example ===

is42 :: Term s (PInteger :--> PBool)
is42 = plam $ i -> unTermCont $ do
  pguardC' (pconstant False) $ i #== 42
  pure $ pconstant True

pletC :: Term s a -> TermCont s (Term s a)

Like plet but works in a TermCont monad

pletFieldsC :: forall fs a s b ps bs. (PDataFields a, ps ~ PFields a, bs ~ Bindings ps fs, BindFields ps bs) => Term s a -> TermCont @b s (HRec (BoundTerms ps bs s))

Like pletFields but works in a TermCont monad.

pmatchC :: PlutusType a => Term s a -> TermCont s (a s)

Like pmatch but works in a TermCont monad

ptraceC :: Term s PString -> TermCont s ()

Like ptrace but works in a TermCont monad.

Example ===

foo :: Term s PUnit
foo = unTermCont $ do
  ptraceC "returning unit!"
  pure $ pconstant ()

ptryFromC :: forall b r a s. PTryFrom a b => Term s a -> TermCont @r s (Term s b, Reduce (PTryFromExcess a b s))

TermCont producing version of ptryFrom.

unTermCont :: TermCont @a s (Term s a) -> Term s a

tcont :: ((a -> Term s r) -> Term s r) -> TermCont @r s a

Tracing

ptrace :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Deprecated: Use ptraceInfo

Backward compatibility synonym for ptraceInfo.

Since: 1.6.0

ptraceDebug :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Trace the given message at the debug level before evaluating the given argument.

Since: 1.6.0

ptraceDebugError :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a

Trace the given message at the debug level, then terminate with perror.

Since: 1.6.0

ptraceDebugIfFalse :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the debug level if the argument is false.

Since: 1.6.0

ptraceDebugIfTrue :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the debug level if the argument is true.

Since: 1.6.0

ptraceDebugShowId :: forall (a :: S -> Type) (s :: S). PShow a => Term s a -> Term s a

Like Haskell's traceShowId but for Plutarch, at the debug level.

Since: 1.6.0

ptraceError :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a

Deprecated: Use ptraceInfoError

Synonym for ptraceInfoError.

Since: 1.6.0

ptraceIfFalse :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Deprecated: Use ptraceInfoIfFalse

Synonym for ptraceInfoIfFalse.

Since: 1.6.0

ptraceInfo :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Trace the given message at the info level before evaluating the given argument.

Since: 1.6.0

ptraceInfoError :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a

Trace the given message at the info level, then terminate with perror.

Since: 1.6.0

ptraceInfoIfFalse :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the info level if the argument is false.

Since: 1.6.0

ptraceInfoIfTrue :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the info level if the argument is true.

Since: 1.6.0

ptraceInfoShowId :: forall (a :: S -> Type) (s :: S). PShow a => Term s a -> Term s a

Like Haskell's traceShowId but for Plutarch, at the info level.

Since: 1.6.0

ptraceShowId :: forall (a :: S -> Type) (s :: S). PShow a => Term s a -> Term s a

Deprecated: Use ptraceInfoShowId

Synonym for ptraceInfoShowId.

Since: 1.6.0