Plutarch.Pretty.Internal.Types

Documentation

data PrettyCursor

Notifies the prettifier what "state" the cursor currently is, so it can decide whether or not to wrap the target expression in parens.

Normal indicates no parens wrapping is necessary, even for complex expressions.

Special indicates complex expressions should be wrapped in parens.

Usually, Special just hints at one of three states:

Applying - The expression is being applied like a function.

Applied - The expression is being applied as a function argument.

Unary arg - The expression is being used as an argument to a high arity unary operator (~ and !).

Constructors

Normal

Special

data PrettyState

Constructors

PrettyState

Fields

ps'nameMap :: Map Index Text

ps'names :: Set Text

ps'cursor :: PrettyCursor

type PrettyMonad s = ReaderT (STGenM StdGen s) (StateT PrettyState (ST s))

forkState :: MonadState s m => m b -> m b

normalizeCursor :: PrettyState -> PrettyState

specializeCursor :: PrettyState -> PrettyState

memorizeName :: Text -> PrettyState -> PrettyState

insertName :: Text -> PrettyState -> PrettyState

Insert a fresh binding onto the name map, i.e a name at index 0 - incrementing all other indices.

insertBindings :: [Text] -> PrettyState -> PrettyState

builtinFunAtRef :: Map Index Text -> Index -> Maybe DefaultFun

nameOfRef :: Index -> Map Index Text -> Maybe Text