Plutarch.Repr.Internal

Documentation

newtype PStruct (struct :: [[S -> Type]]) (s :: S)

Since: 1.10.0

Constructors

PStruct

Fields

unPStruct :: SOP (Term s) struct

Since: 1.10.0

newtype PRec (struct :: [S -> Type]) (s :: S)

Since: 1.10.0

Constructors

PRec

Fields

unPRec :: NP (Term s) struct

Since: 1.10.0

pletL :: All SListI as => SOP (Term s) as -> (SOP (Term s) as -> Term s r) -> Term s r

Since: 1.10.0

grecEq :: forall (s :: S) (struct :: [S -> Type]). All PEq struct => NP (Term s) struct -> NP (Term s) struct -> Term s PBool

Since: 1.10.0

gstructEq :: forall (s :: S) (struct :: [[S -> Type]]). All2 PEq struct => SOP (Term s) struct -> SOP (Term s) struct -> Term s PBool

Since: 1.10.0

groupHandlers :: forall (s :: S) (r :: S -> Type). [(Integer, Term s r)] -> Term s PInteger -> Term s r

This function handles optimization of function that require multiple handlers by checking hashes of each | handler item and merging them in a way it will minimize size and cost of all computation

Since: 1.10.0

class (Generic (a s), AllZipN @Type (Prod SOP) (LiftedCoercible I (Term s)) (Code (a s)) struct, AllZipN @Type (Prod SOP) (LiftedCoercible (Term s) I) struct (Code (a s))) => StructSameRepr s a struct

Since: 1.10.0

type family UnTermRec (struct :: [Type]) :: [S -> Type] where ...

Since: 1.10.0

Equations

UnTermRec '[] = '[]

UnTermRec (Term _ a ': rest) = a ': UnTermRec rest

type UnTermStruct x = UnTermStruct' (Code x)

Since: 1.10.0

type RecTypePrettyError struct = RecTypePrettyError' struct ~ 'True

Since: 1.10.0