Plutarch.Repr.SOP

Documentation

newtype PSOPStruct (struct :: [[S -> Type]]) (s :: S)

@since WIP

Constructors

PSOPStruct

Fields

unPSOPStruct :: PStruct struct s

@since WIP

newtype PSOPRec (struct :: [S -> Type]) (s :: S)

@since WIP

Constructors

PSOPRec

Fields

unPSOPRec :: PRec struct s

@since WIP

newtype DeriveAsSOPStruct (a :: S -> Type) s

via-derivation helper to derive PlutusType instance using SoP encoding. If your type has only one constructor prefer using DeriveAsSOPRec instead.

@since WIP

Constructors

DeriveAsSOPStruct

Fields

unDeriveAsSOPStruct :: a s

@since WIP

newtype DeriveAsSOPRec (a :: S -> Type) s

via-derivation helper for SoP encoding, currently behaves exactly like DeriveAsSOPStruct but can be used only on types with a single constructor. It is separate to leave a room for future optimizations

@since WIP

Constructors

DeriveAsSOPRec

Fields

unDeriveAsSOPRec :: a s

@since WIP