Plutarch.TermCont

Documentation

hashOpenTerm :: Term s a -> TermCont s Dig

newtype TermCont :: forall (r :: PType). S -> Type -> Type where

Constructors

TermCont :: forall r s a. ((a -> Term s r) -> Term s r) -> TermCont @r s a

$sel:runTermCont:TermCont :: TermCont @r s a -> (a -> Term s r) -> Term s r

unTermCont :: TermCont @a s (Term s a) -> Term s a

tcont :: ((a -> Term s r) -> Term s r) -> TermCont @r s a

pletC :: Term s a -> TermCont s (Term s a)

Like plet but works in a TermCont monad

pmatchC :: PlutusType a => Term s a -> TermCont s (a s)

Like pmatch but works in a TermCont monad

pletFieldsC :: forall fs a s b ps bs. (PDataFields a, ps ~ PFields a, bs ~ Bindings ps fs, BindFields ps bs) => Term s a -> TermCont @b s (HRec (BoundTerms ps bs s))

Like pletFields but works in a TermCont monad.

ptraceC :: Term s PString -> TermCont s ()

Like ptrace but works in a TermCont monad.

Example ===

foo :: Term s PUnit
foo = unTermCont $ do
  ptraceC "returning unit!"
  pure $ pconstant ()

pguardC :: Term s PString -> Term s PBool -> TermCont s ()

Trace a message and raise error if cond is false. Otherwise, continue.

Example ===

onlyAllow42 :: Term s (PInteger :--> PUnit)
onlyAllow42 = plam $ i -> unTermCont $ do
  pguardC "expected 42" $ i #== 42
  pure $ pconstant ()

pguardC' :: Term s a -> Term s PBool -> TermCont @a s ()

Stop computation and return given term if cond is false. Otherwise, continue.

Example ===

is42 :: Term s (PInteger :--> PBool)
is42 = plam $ i -> unTermCont $ do
  pguardC' (pconstant False) $ i #== 42
  pure $ pconstant True

ptryFromC :: forall b r a s. PTryFrom a b => Term s a -> TermCont @r s (Term s b, Reduce (PTryFromExcess a b s))

TermCont producing version of ptryFrom.

pexpectJustC :: forall (a :: S -> Type) (r :: S -> Type) (s :: S). Term s r -> Term s (PMaybe a) -> TermCont @r s (Term s a)

Escape with a particular value on expecting PJust. For use in monadic context.

@since WIP