Plutarch.Trace

Documentation

ptrace :: Term s PString -> Term s a -> Term s a

Trace the given message before evaluating the argument.

ptraceShowId :: PShow a => Term s a -> Term s a

Like Haskell's traceShowId but for Plutarch

ptraceIfTrue :: Term s PString -> Term s PBool -> Term s PBool

Trace the given message if the argument evaluates to true.

ptraceIfFalse :: Term s PString -> Term s PBool -> Term s PBool

Trace the given message if the argument evaluates to False.

ptraceError :: Term s PString -> Term s a

Trace the given message and terminate evaluation with a perror.