Plutarch.Unsafe

Documentation

punsafeBuiltin :: DefaultFun -> Term s a

punsafeCoerce :: Term s a -> Term s b

Unsafely coerce the type-tag of a Term.

This should mostly be avoided, though it can be safely used to assert known types of Datums, Redeemers or ScriptContext.

punsafeConstant :: Some (ValueOf DefaultUni) -> Term s a

Deprecated: Use pconstant instead.

punsafeDowncast :: Term s (PInner a) -> Term s a

Unsafely coerce from the PInner representation of a Term, assuming that the value is a safe construction of the Term.