Plutarch

Documentation

data ((a :: PType) :--> (b :: PType)) (s :: S) infixr 0

type ClosedTerm (a :: PType) = forall (s :: S). Term s a

Closed* terms with no free variables.

compile :: Config -> ClosedTerm a -> Either Text Script

Compile a (closed) Plutus Term to a usable script

newtype Script

Constructors

Script (Program DeBruijn DefaultUni DefaultFun ())

type Dig = Digest Blake2b_160

hashTerm :: Config -> ClosedTerm a -> Either Text Dig

papp :: Term s (a :--> b) -> Term s a -> Term s b

Lambda Application.

pdelay :: Term s a -> Term s (PDelayed a)

Plutus 'delay', used for laziness.

data PDelayed (a :: PType) (s :: S)

perror :: Term s a

Plutus 'error'.

When using this explicitly, it should be ensured that the containing term is delayed, avoiding premature evaluation.

pforce :: Term s (PDelayed a) -> Term s a

Plutus 'force', used to force evaluation of PDelayed terms.

phoistAcyclic :: HasCallStack => ClosedTerm a -> Term s a

plet :: Term s a -> (Term s a -> Term s b) -> Term s b

Let bindings.

This is approximately a shorthand for a lambda and application:

plet v f == papp (plam f) v

But sufficiently small terms in WHNF may be inlined for efficiency.

pthrow :: HasCallStack => Text -> Term s a

data Term (s :: S) (a :: PType)

data S

type PType = S -> Type

Shorthand for Plutarch types.

(#$) :: Term s (a :--> b) -> Term s a -> Term s b infixr 0

Low precedence infixr synonym of papp, to be used like $, in combination with #. e.g.:

>>> f # x #$ g # y # z
f x (g y z)

(#) :: Term s (a :--> b) -> Term s a -> Term s b infixl 8

High precedence infixl synonym of papp, to be used like function juxtaposition. e.g.:

>>> f # x # y
f x y

class PlutusType (a :: PType)

type family PInner a :: PType

pcon :: PlutusType a => a s -> Term s a

Construct a Plutarch Term via a Haskell datatype

pmatch :: PlutusType a => Term s a -> (a s -> Term s b) -> Term s b

Pattern match over Plutarch Terms via a Haskell datatype

type PCon = PlutusType

Deprecated: Use PlutusType

type PMatch = PlutusType

Deprecated: Use PlutusType

prettyTerm :: Config -> ClosedTerm a -> Doc ()

Prettify a Plutarch term.

This will call error if there's a compilation failure. Use prettyTerm' for a non-partial version.

Example ==

import Plutarch.Prelude
import Plutarch.Api.V1

checkSignatory :: Term s (PPubKeyHash :--> PScriptContext :--> PUnit)
checkSignatory = plam $ ph ctx' -> unTermCont $ do
  ctx <- pletFieldsC ["txInfo", "purpose"] ctx'
  purph <- pmatchC ctx.purpose
  pure $ case purph of
    PSpending _ ->
      let signatories = pfield "signatories" # ctx.txInfo
      in pif
          (pelem # pdata ph # pfromData signatories)
          -- Success!
          (pconstant ())
          -- Signature not present.
          perror
    _ -> ptraceError "checkSignatoryCont: not a spending tx"

Prettification result:

let frSndPair = !!sndPair
    unDataSum = (xF -> frSndPair (unConstrData xF))
    frTailList = !tailList
    frHeadList = !headList
    frIfThenElse = !ifThenElse
in (oP4ECBT qsrxlF0Y7 ->
      let cjlB6yrGk = unDataSum qsrxlF0Y7
          cRFO = unConstrData (frHeadList (frTailList cjlB6yrGk))
          cs9iR = !!fstPair cRFO
          w4 = frSndPair cRFO
      in if equalsInteger 1 cs9iR
           then if (vModHwqYB ->
                      let blM6d67 =
                            (x5sad ePDSInSEC ->
                               !(!!chooseList
                                   ePDSInSEC
                                   ~False
                                   ~(if equalsData
                                          (frHeadList ePDSInSEC)
                                          vModHwqYB
                                       then True
                                       else x5sad (frTailList ePDSInSEC))))
                          mC = (jfZs -> blM6d67 (itzT -> jfZs jfZs itzT))
                      in blM6d67 (ispwp_oeT -> mC mC ispwp_oeT))
                     (bData oP4ECBT)
                     (unListData
                        let q6X3 = frHeadList cjlB6yrGk
                        in frHeadList
                             let olbZ = unDataSum q6X3
                             in frTailList
                                  (frTailList
                                     (frTailList
                                        (frTailList
                                           (frTailList
                                              (frTailList
                                                 (frTailList olbZ)))))))
                  then ()
                  else ERROR
           else !(!trace "checkSignatoryCont: not a spending tx" ~ERROR))

Semantics ==

Constants ===

Builtin integers are printed as regular integers. [0-9]+

Builtin bytestrings are printed in hex notation, prefixed by `0x`. 0x[0-9a-f]+/i

Builtin strings are printed as is.

Builtin unit is printed as the unit literal. ()

Builtin booleans are printed as the literal True or False.

Builtin lists are prettified as list literals, i.e delimited with `[` and `]`.

Builtin pairs are prettified as 2-ary tuple literals, e.g. `(a, b)`.

I data (i.e data encoded integers) are prettified like builtin integers with a # prefix. #[0-9]+

B data (i.e data encoded bytestrings) are prettified like builtin bytestrings with a # prefix. #0x[0-9a-f]+/i

List data (i.e data encoded lists) are prettified like builtin lists with a # prefix.

Map data is printed like record literals. Delimited by `{` and `}`.

Each key value pair is prettified like key = value and multiple pairs are joined with `,`.

For example, `Map [(I 42, I 0), (I 100, I 1)]` is prettified as `{ #42 = #0, #100 = #1 }` - Constr data has two core elements in its prettified form:

The constructor index, prettified as an integer prefixed with Σ (sigma).

Its fields, prettified as a list.

These two elements are then joined with a . (period).

For example, `Constr 1 [I 42]` is prettified as "Σ1.[#42]".

Builtin functions ===

Builtin functions are prettified into their name, in title case.

Forced term ===

Forced terms are prefixed with a !. The unary operator ! has higher fixity than function application.

Delayed term ===

Delayed terms are prefixed with a ~. The unary operator ~ has higher fixity than function application.

Var ===

Random names are generated for all variable bindings, and these names are used to refer to them.

Names are always unique, between 1 and 8 characters in length, and begin with a lowercase letter.

Names may consist of alphanumeric characters, underscore, or single quotes.

LamAbs ===

Lambdas are prettified similar to haskell lambdas, i.e `x -> ...`.

Lambdas with multiple arguments are detected and simplified: `x y z -> ...`.

Apply ===

Application is, simply, a space - just like haskell. `f x`.

Multi arg applications to the same function are detected and simplified: `f x y`.

Error term ===

perror is represented by the literal ERROR.

Special handling ===

To achieve better prettification, certain AST structures are given special handling logic.

The AST structure produced by plet (Single Apply + LamAbs pair) is prettified into Haskell-like let bindings.

Lazy ifthenelse (pif in particular, not pif') is detected and prettified into Haskell-like syntax: `if cond then expr1 else expr2`.

Chains of ifthenelse are nested:

 if cond
   then expr1
   else if cond
     then expr2
     else expr3
 

When generating names for bindings, well known structures are identified and given special names.

This machinery is made to be extensible in the future.

For example, the structure of the pfix function is well known and constant - so it is simply called fix in the output.

Bindings to forced builtin functions inherit the builtin function name, prefixed with a fr.

prettyScript :: Script -> Doc ()

prettyTerm for pre-compiled Scripts.

printTerm :: HasCallStack => Config -> ClosedTerm a -> String

Prettyprint a Term via the PLC pretty printer

TODO: Heavily improve. It's unreadable right now.

We could convert the de Bruijn indices into names with:

show . prettyPlcReadableDef . (\(Right p) -> p) . Scripts.mkTermToEvaluate . compile $ term

printScript :: Script -> String

Prettyprint a compiled Script via the PLC pretty printer

pinl :: Term s a -> (Term s a -> Term s b) -> Term s b

pto :: Term s a -> Term s (PInner a)

Safely coerce from a Term to it's PInner representation.

pfix :: Term s (((a :--> b) :--> (a :--> b)) :--> (a :--> b))

Fixpoint recursion. Used to encode recursive functions.

Example:

iterateN' ::
 Term s (PInteger :--> (a :--> a) :--> a :--> a) ->
 Term s PInteger ->
 Term s (a :--> a) ->
 Term s a
iterateN' self n f x =
  pif (n #== 0) x (self # n - 1 #$ f x)

iterateN :: Term s (PInteger :--> (a :--> a) :--> a :--> a)
iterateN = pfix #$ plam iterateN'

Further examples can be found in examples/Recursion.hs

newtype POpaque s

An Arbitrary Term with an unknown type

Constructors

POpaque (Term s POpaque)

popaque :: Term s a -> Term s POpaque

Erase the type of a Term

plam :: forall c. (PLamN a b s, HasCallStack) => (Term s c -> a) -> Term s (c :--> b)

newtype TermCont :: forall (r :: PType). S -> Type -> Type where

Constructors

TermCont :: forall r s a. ((a -> Term s r) -> Term s r) -> TermCont @r s a

hashOpenTerm :: Term s a -> TermCont s Dig

$sel:runTermCont:TermCont :: TermCont @r s a -> (a -> Term s r) -> Term s r

unTermCont :: TermCont @a s (Term s a) -> Term s a

tcont :: ((a -> Term s r) -> Term s r) -> TermCont @r s a

data Config where

Configuration for Plutarch scripts at compile time. This indicates whether we want to trace, and if so, under what log level and mode.

Since: 1.6.0

Bundled Patterns

pattern Tracing :: LogLevel -> TracingMode -> Config
Pattern for a tracing config, with both its log level and mode.

Since: 1.6.0

pattern NoTracing :: Config
Pattern for the config that does no tracing (also the default).

Since: 1.6.0

data TracingMode

How to trace.

Since: 1.6.0

Constructors

DetTracing

DoTracing

DoTracingAndBinds

data LogLevel

What logging level we want to use.

Since: 1.6.0

Constructors

LogInfo

LogDebug

tracingMode :: Config -> Maybe TracingMode

If the config indicates that we want to trace, get its mode.

Since: 1.6.0

logLevel :: Config -> Maybe LogLevel

If the config indicates that we want to trace, get its log level.

Since: 1.6.0

pgetConfig :: (Config -> Term s a) -> Term s a

newtype PForall (b :: a -> PType) s

Constructors

PForall (forall (x :: a). Term s (b x))

data PSome (b :: a -> PType) s

Constructors

forall (x :: a). PSome (Term s (b x))

newtype PScottEncoded a r s

Constructors

PScottEncoded (Term s (ScottFn (ScottList a r) r))

data PlutusTypeScott

data PlutusTypeNewtype

class (PInner a ~ DerivedPInner (DPTStrat a) a, PlutusTypeStrat (DPTStrat a), PlutusTypeStratConstraint (DPTStrat a) a, PlutusType a) => DerivePlutusType (a :: PType)

type family DPTStrat a :: Type

class (forall t. PCovariant'' t => PCovariant'' (a t)) => PCovariant a

type family PCovariant' a :: Constraint

class (forall t. PCovariant'' t => PContravariant'' (a t)) => PContravariant a

type family PContravariant' a :: Constraint

class (forall t. PVariant'' t => PVariant'' (a t)) => PVariant a

type family PVariant' a :: Constraint