Plutarch.Api.V1.AssocMap

Documentation

newtype PMap (keysort :: KeyGuarantees) (k :: PType) (v :: PType) (s :: S)

Constructors

PMap (Term s (PBuiltinList (PBuiltinPair (PAsData k) (PAsData v))))

data KeyGuarantees

Constructors

Sorted

Unsorted

Creation

pempty :: Term s (PMap 'Sorted k v)

Construct an empty PMap.

psingleton :: (PIsData k, PIsData v) => Term s (k :--> (v :--> PMap 'Sorted k v))

Construct a singleton PMap with the given key and value.

psingletonData :: Term s (PAsData k :--> (PAsData v :--> PMap 'Sorted k v))

Construct a singleton PMap with the given data-encoded key and value.

pinsert :: (POrd k, PIsData k, PIsData v) => Term s (k :--> (v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Insert a new key/value pair into the map, overriding the previous if any.

pinsertData :: (POrd k, PIsData k) => Term s (PAsData k :--> (PAsData v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Insert a new data-encoded key/value pair into the map, overriding the previous if any.

pdelete :: (POrd k, PIsData k) => Term s (k :--> (PMap 'Sorted k v :--> PMap 'Sorted k v))

Delete a key from the map.

pfromAscList :: (POrd k, PIsData k, PIsData v) => Term s (PBuiltinListOfPairs k v :--> PMap 'Sorted k v)

Construct a PMap from a list of key-value pairs, sorted by ascending key data.

passertSorted :: forall k v any s. (POrd k, PIsData k, PIsData v) => Term s (PMap any k v :--> PMap 'Sorted k v)

Assert the map is properly sorted.

pforgetSorted :: Term s (PMap 'Sorted k v) -> Term s (PMap g k v)

Forget the knowledge that keys were sorted.

Lookups

plookup :: (PIsData k, PIsData v) => Term s (k :--> (PMap any k v :--> PMaybe v))

Look up the given key in a PMap.

plookupData :: Term s (PAsData k :--> (PMap any k v :--> PMaybe (PAsData v)))

Look up the given key data in a PMap.

pfindWithDefault :: (PIsData k, PIsData v) => Term s (v :--> (k :--> (PMap any k v :--> v)))

Look up the given key in a PMap, returning the default value if the key is absent.

pfoldAt :: PIsData k => Term s (k :--> (r :--> ((PAsData v :--> r) :--> (PMap any k v :--> r))))

Look up the given key in a PMap; return the default if the key is absent or apply the argument function to the value data if present.

pnull :: Term s (PMap any k v :--> PBool)

Tests whether the map is empty.

Folds

pall :: PIsData v => Term s ((v :--> PBool) :--> (PMap any k v :--> PBool))

Tests if all values in the map satisfy the given predicate.

pany :: PIsData v => Term s ((v :--> PBool) :--> (PMap any k v :--> PBool))

Tests if anu value in the map satisfies the given predicate.

Filters and traversals

pfilter :: PIsData v => Term s ((v :--> PBool) :--> (PMap g k v :--> PMap g k v))

Filters the map so it contains only the values that satisfy the given predicate.

pmap :: (PIsData a, PIsData b) => Term s ((a :--> b) :--> (PMap g k a :--> PMap g k b))

Applies a function to every value in the map, much like map.

pmapData :: Term s ((PAsData a :--> PAsData b) :--> (PMap g k a :--> PMap g k b))

pmapMaybe :: (PIsData a, PIsData b) => Term s ((a :--> PMaybe b) :--> (PMap g k a :--> PMap g k b))

Maps and filters the map, much like mapMaybe.

pmapMaybeData :: Term s ((PAsData a :--> PMaybe (PAsData b)) :--> (PMap g k a :--> PMap g k b))

Combining

type BothPresentHandler k v s = BothPresentHandler_ (->) (Term s k) (Term s v)

data BothPresentHandler_ f k v

Constructors

DropBoth

PassArg Bool
True ~ left, False ~ right

HandleBoth (k `f` (v `f` (v `f` v)))

type BothPresentHandlerCommutative k v s = BothPresentHandlerCommutative_ (->) (Term s k) (Term s v)

data BothPresentHandlerCommutative_ f k v

Constructors

DropBothCommutative

HandleBothCommutative (k `f` (v `f` (v `f` v)))

type OnePresentHandler k v s = OnePresentHandler_ (->) (Term s k) (Term s v)

data OnePresentHandler_ f k v

Constructors

DropOne

PassOne

HandleOne (k `f` (v `f` v))

type MergeHandler k v s = MergeHandler_ (->) (Term s k) (Term s v)

data MergeHandler_ f k v

Signals how to handle value merging for matching keys in pzipWith.

No restrictions on commutativity: Safe to use for both commutative and non-commutative merging operations.

Constructors

MergeHandler

Fields

mhBoth :: BothPresentHandler_ f k v

mhLeft :: OnePresentHandler_ f k v

mhRight :: OnePresentHandler_ f k v

type MergeHandlerCommutative k v s = MergeHandlerCommutative_ (->) (Term s k) (Term s v)

data MergeHandlerCommutative_ f k v

Signals how to handle value merging for matching keys in pzipWith.

Safe to use for commutative merging operations only.

Constructors

MergeHandlerCommutative

Fields

mhcBoth :: BothPresentHandlerCommutative_ f k v

mhcOne :: OnePresentHandler_ f k v

type SomeMergeHandler k v s = SomeMergeHandler_ (->) (Term s k) (Term s v)

data SomeMergeHandler_ f k v

Signals how to handle value merging for matching keys in pzipWith.

Constructors

SomeMergeHandlerCommutative (MergeHandlerCommutative_ f k v)

SomeMergeHandler (MergeHandler_ f k v)

data Commutativity

Constructors

Commutative

NonCommutative

pzipWith :: forall (s :: S) (k :: PType) (v :: PType). (POrd k, PIsData k, PIsData v) => SomeMergeHandler k v s -> Term s (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v))

Zip two PMaps, using a SomeMergeHandler to decide how to merge based on key presence on the left and right.

Note that using a MergeHandlerCommutative is less costly than a MergeHandler.

pzipWithData :: forall (s :: S) (k :: PType) (v :: PType). (POrd k, PIsData k) => SomeMergeHandler (PAsData k) (PAsData v) s -> Term s (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v))

Zip two PMaps, using a SomeMergeHandler to decide how to merge based on key presence on the left and right.

Note that using a MergeHandlerCommutative is less costly than a MergeHandler.

pzipWithDefault :: forall (s :: S) (k :: PType) (v :: PType). (POrd k, PIsData k, PIsData v) => ClosedTerm v -> Commutativity -> Term s ((v :--> (v :--> v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Zip two PMaps, using the given value merge function for key collisions, the commutativity of the merge function, and a default value that can stand in for both sides.

Note that using Commutative is less costly than NonCommutative.

pzipWithDataDefault :: forall (s :: S) (k :: PType) (v :: PType). (POrd k, PIsData k) => ClosedTerm (PAsData v) -> Commutativity -> Term s ((PAsData v :--> (PAsData v :--> PAsData v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Zip two PMaps, using the given value merge function for key collisions, the commutativity of the merge function, and a default value that can stand in for both sides.

Note that using Commutative is less costly than NonCommutative.

pzipWithDefaults :: forall (s :: S) (k :: PType) (v :: PType). (POrd k, PIsData k, PIsData v) => ClosedTerm v -> ClosedTerm v -> Term s ((v :--> (v :--> v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Zip two PMaps, using the given potentially non-commutative value merge function for key collisions, and different values for the sides.

Use pzipWithDefault if your merge function is commutative for better performance.

pzipWithDataDefaults :: forall (s :: S) (k :: PType) (v :: PType). (POrd k, PIsData k) => ClosedTerm (PAsData v) -> ClosedTerm (PAsData v) -> Term s ((PAsData v :--> (PAsData v :--> PAsData v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Zip two PMaps, using the given potentially non-commutative value merge function for key collisions, and different values for the sides.

Use pzipWithDataDefault if your merge function is commutative for better performance.

pintersectionWith :: (POrd k, PIsData k, PIsData v) => Commutativity -> Term s ((v :--> (v :--> v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Build the intersection of two PMaps, merging values that share the same key using the given function.

pintersectionWithData :: (POrd k, PIsData k) => Commutativity -> Term s ((PAsData v :--> (PAsData v :--> PAsData v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Build the intersection of two PMaps, merging data-encoded values that share the same key using the given function.

pleftBiasedUnion :: (POrd k, PIsData k, PIsData v) => Term s (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v))

Build the union of two PMaps. Take the value from the left argument for colliding keys.

Prefer this over 'punionResolvingCollisionsWith NonCommutative # plam const'. It performs better.

punionResolvingCollisionsWith :: (POrd k, PIsData k, PIsData v) => Commutativity -> Term s ((v :--> (v :--> v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Build the union of two PMaps, merging values that share the same key using the given function.

punionResolvingCollisionsWithData :: (POrd k, PIsData k) => Commutativity -> Term s ((PAsData v :--> (PAsData v :--> PAsData v)) :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v)))

Build the union of two PMaps, merging values that share the same key using the given function.

pdifference :: (PIsData k, POrd k, PIsData v) => Term s (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PMap 'Sorted k v))

Difference of two maps. Return elements of the first map not existing in the second map.

punsortedDifference :: PIsData k => Term s (PMap g k a :--> (PMap any k b :--> PMap g k a))

Difference of two maps. Return elements of the first map not existing in the second map. Warning: O(n^2).

Partial order operations

pcheckBinRel :: forall k v s. (POrd k, PIsData k, PIsData v) => Term s ((v :--> (v :--> PBool)) :--> (v :--> (PMap 'Sorted k v :--> (PMap 'Sorted k v :--> PBool))))

Given a comparison function and a "zero" value, check whether a binary relation holds over 2 sorted PMaps.

This is primarily intended to be used with PValue.

Orphan instances

(PConstantData k, PConstantData v, Ord k) => PConstantDecl (Map k v)

Instance details

Associated Types

type PConstantRepr (Map k v)

type PConstanted (Map k v) :: PType

Methods

pconstantToRepr :: Map k v -> PConstantRepr (Map k v)

pconstantFromRepr :: PConstantRepr (Map k v) -> Maybe (Map k v)