Plutarch.BitString

Type

newtype PBitString (s :: S)

A wrapper around PByteString for CIP-122 and CIP-123 bitwise operations.

Note

This type exists because bit and byte indexes work in different directions. To avoid confusing behaviour, we require an explicit wrapping of PByteStrings to use bitwise functionality: this way, it's clear where which scheme applies.

@since WIP

Constructors

PBitString (Term s PByteString)

Functions

preadBit :: forall (s :: S). Term s (PBitString :--> (PInteger :--> PBool))

Bit access operation, as defined in CIP-122.

@since WIP

pwriteBits :: forall (s :: S). Term s (PBitString :--> (PBuiltinList PInteger :--> (PBool :--> PBitString)))

Given a list of positions, set the bits at those positions.

This works similarly to the writeBits operation in CIP-122 with regard to the list of indexes. Effectively, psetBits b ixes is equivalent to writeBits b . map (, True) $ ixes. All caveats that this entails from the CIP-122 description apply.

@since WIP

Sets bits, as per CIP-122.

@since WIP

pshift :: forall (s :: S). Term s (PBitString :--> (PInteger :--> PBitString))

Performs a shift, as per CIP-123.

@since WIP

protate :: forall (s :: S). Term s (PBitString :--> (PInteger :--> PBitString))

Performs a rotation, as per CIP-123.

@since WIP

pcountSetBits :: forall (s :: S). Term s (PBitString :--> PInteger)

Counts the number of set bits, as per CIP-123.

@since WIP

pfindFirstSetBit :: forall (s :: S). Term s (PBitString :--> PMaybe PInteger)

As pfindFirstSetBit', but produces PNothing if the argument is empty, or contains no set bits.

@since WIP

pfindFirstSetBit' :: forall (s :: S). Term s (PBitString :--> PInteger)

Finds the index of the first set bit, as per CIP-123.

Note

This returns -1 if the argument is either empty, or contains no set bits.

@since WIP