Plutarch.DataRepr.Internal

Documentation

newtype PDataSum defs s

A sum of PDataRecords. The underlying representation is the Constr constructor, where the integer is the index of the variant and the list is the record.

Constructors

PDataSum (NS (Compose (Term s) PDataRecord) defs)

punDataSum :: Term s (PDataSum '[def] :--> PDataRecord def)

If there is only a single variant, then we can safely extract it.

ptryIndexDataSum :: KnownNat n => Proxy n -> Term s (PDataSum (def ': defs) :--> PDataRecord (IndexList n (def ': defs)))

Try getting the nth variant. Errs if it's another variant.

pdcons :: forall label a l s. Term s (PAsData a :--> (PDataRecord l :--> PDataRecord ((label ':= a) ': l)))

Cons a field to a data record.

You can specify the label to associate with the field using type applications-

foo :: Term s (PDataRecord '[ "fooField" ':= PByteString ])
foo = pdcons @"fooField" # pdata (phexByteStr "ab") # pdnil

pdnil :: Term s (PDataRecord '[])

An empty PDataRecord.

data DataReprHandlers (out :: PType) (defs :: [[PLabeledType]]) (s :: S) where

Constructors

DRHNil :: DataReprHandlers out '[] s

DRHCons :: (Term s (PDataRecord def) -> Term s out) -> DataReprHandlers out defs s -> DataReprHandlers out (def ': defs) s

type PConstantData h = (PConstant h, FromData h, ToData h, PIsData (PConstanted h))

Type synonym to simplify deriving of PConstant via DerivePConstantViaData.

A type Foo a is considered ConstantableData if:

The wrapped type a has a PConstant instance.

The lifted type of a has a PUnsafeLiftDecl instance.

There is type equality between a and PLifted (PConstanted a).

The newtype has FromData and ToData instances

These constraints are sufficient to derive a PConstant instance for the newtype.

For deriving PConstant for a wrapped type represented in UPLC as Data, see DerivePConstantViaData.

Polymorphic types can be derived as follows:

data Bar a = Bar a deriving stock (GHC.Generic)

PlutusTx.makeLift ''Bar
PlutusTx.makeIsDataIndexed ''Bar [('Bar, 0)]

data PBar (a :: PType) (s :: S)
  = PBar (Term s (PDataRecord '["_0" ':= a]))
  deriving stock (GHC.Generic)
  deriving anyclass (SOP.Generic, PIsDataRepr)
  deriving (PlutusType, PIsData, PDataFields) via PIsDataReprInstances (PBar a)

instance
  forall a.
  PLiftData a =>
  PUnsafeLiftDecl (PBar a)
  where
  type PLifted (PBar a) = Bar (PLifted a)

deriving via
  ( DerivePConstantViaData
      (Bar a)
      (PBar (PConstanted a))
  )
  instance
    PConstantData a =>
    PConstantDecl (Bar a)

data PDataRecord (as :: [PLabeledType]) (s :: S) where

A "record" of `exists a. PAsData a`. The underlying representation is `PBuiltinList PData`.

Constructors

PDCons :: forall name_x x xs s. PUnLabel name_x ~ x => Term s (PAsData x) -> Term s (PDataRecord xs) -> PDataRecord (name_x ': xs) s

PDNil :: PDataRecord '[] s

type PLiftData p = (PLift p, FromData (PLifted p), ToData (PLifted p), PIsData p)

data PLabeledType

Constructors

Symbol := PType

type family PLabelIndex (name :: Symbol) (as :: [PLabeledType]) :: Nat where ...

Equations

PLabelIndex name ((name ':= _) ': _) = 0

PLabelIndex name (_ ': as) = PLabelIndex name as + 1

type family PUnLabel (a :: PLabeledType) :: PType where ...

Equations

PUnLabel (_ ':= a) = a

type family PLookupLabel name as where ...

Equations

PLookupLabel name ((name ':= a) ': _) = a

PLookupLabel name (_ ': as) = PLookupLabel name as

pindexDataRecord :: KnownNat n => Proxy n -> Term s (PDataRecord as) -> Term s (PAsData (PUnLabel (IndexList n as)))

Safely index a PDataRecord.

pdropDataRecord :: KnownNat n => Proxy n -> Term s (PDataRecord xs) -> Term s (PDataRecord (Drop n xs))

Safely drop the first n items of a PDataRecord.

newtype DerivePConstantViaData (h :: Type) (p :: PType)

For deriving PConstant for a wrapped type represented by a builtin type, see DerivePConstantViaNewtype.

Constructors

DerivePConstantViaData h
The Haskell type for which @PConstant is being derived.

newtype DualReprHandler s out def

Constructors

DualRepr (Term s (PDataRecord def) -> Term s (PDataRecord def) -> Term s out)

data PlutusTypeData