Plutarch.Evaluate

Documentation

evalScript :: Script -> (Either EvalError Script, ExBudget, [Text])

Evaluate a script with a big budget, returning the trace log and term result.

evalScriptHuge :: Script -> (Either EvalError Script, ExBudget, [Text])

Evaluate a script with a huge budget, returning the trace log and term result.

evalScript' :: ExBudget -> Script -> (Either (CekEvaluationException NamedDeBruijn DefaultUni DefaultFun) Script, ExBudget, [Text])

Evaluate a script with a specific budget, returning the trace log and term result.

type EvalError = CekEvaluationException NamedDeBruijn DefaultUni DefaultFun