Plutarch.Internal

Documentation

$hoisted

newtype ((a :: PType) :--> (b :: PType)) (s :: S) infixr 0

Constructors

PLam (Term s a -> Term s b)

data PDelayed (a :: PType) (s :: S)

$term

newtype Term (s :: S) (a :: PType)

Constructors

Term

Fields

asRawTerm :: Word64 -> TermMonad TermResult

asClosedRawTerm :: ClosedTerm a -> TermMonad TermResult

newtype Script

Constructors

Script (Program DeBruijn DefaultUni DefaultFun ())

mapTerm :: (RawTerm -> RawTerm) -> TermResult -> TermResult

plam' :: (Term s a -> Term s b) -> Term s (a :--> b)

Lambda abstraction.

Only works with a single argument. Use plam instead, to support currying.

plet :: Term s a -> (Term s a -> Term s b) -> Term s b

Let bindings.

This is approximately a shorthand for a lambda and application:

plet v f == papp (plam f) v

But sufficiently small terms in WHNF may be inlined for efficiency.

papp :: Term s (a :--> b) -> Term s a -> Term s b

Lambda Application.

pdelay :: Term s a -> Term s (PDelayed a)

Plutus 'delay', used for laziness.

pforce :: Term s (PDelayed a) -> Term s a

Plutus 'force', used to force evaluation of PDelayed terms.

phoistAcyclic :: HasCallStack => ClosedTerm a -> Term s a

perror :: Term s a

Plutus 'error'.

When using this explicitly, it should be ensured that the containing term is delayed, avoiding premature evaluation.

punsafeCoerce :: Term s a -> Term s b

Unsafely coerce the type-tag of a Term.

This should mostly be avoided, though it can be safely used to assert known types of Datums, Redeemers or ScriptContext.

punsafeBuiltin :: DefaultFun -> Term s a

punsafeConstant :: Some (ValueOf DefaultUni) -> Term s a

Deprecated: Use pconstant instead.

punsafeConstantInternal :: Some (ValueOf DefaultUni) -> Term s a

compile :: Config -> ClosedTerm a -> Either Text Script

Compile a (closed) Plutus Term to a usable script

compileOptimized :: forall (a :: S -> Type). (forall (s :: S). Term s a) -> Either Text Script

As compile, but performs UPLC optimizations. Furthermore, this will always elide tracing (as if with NoTracing).

@since WIP

compile' :: TermResult -> UTerm

type ClosedTerm (a :: PType) = forall (s :: S). Term s a

Closed* terms with no free variables.

type Dig = Digest Blake2b_160

hashTerm :: Config -> ClosedTerm a -> Either Text Dig

hashRawTerm :: RawTerm -> Dig

data RawTerm

Constructors

RVar Word64

RLamAbs Word64 RawTerm

RApply RawTerm [RawTerm]

RForce RawTerm

RDelay RawTerm

RConstant (Some (ValueOf DefaultUni))

RBuiltin DefaultFun

RCompiled UTerm

RError

RHoisted HoistedTerm

data TermResult

Constructors

TermResult

Fields

getTerm :: RawTerm

getDeps :: [HoistedTerm]

data S

Constructors

SI

type PType = S -> Type

Shorthand for Plutarch types.

pthrow :: HasCallStack => Text -> Term s a

data Config where

Configuration for Plutarch scripts at compile time. This indicates whether we want to trace, and if so, under what log level and mode.

Since: 1.6.0

Bundled Patterns

pattern NoTracing :: Config
Pattern for the config that does no tracing (also the default).

Since: 1.6.0

pattern Tracing :: LogLevel -> TracingMode -> Config
Pattern for a tracing config, with both its log level and mode.

Since: 1.6.0

data TracingMode

How to trace.

Since: 1.6.0

Constructors

DetTracing

DoTracing

DoTracingAndBinds

data LogLevel

What logging level we want to use.

Since: 1.6.0

Constructors

LogInfo

LogDebug

tracingMode :: Config -> Maybe TracingMode

If the config indicates that we want to trace, get its mode.

Since: 1.6.0

logLevel :: Config -> Maybe LogLevel

If the config indicates that we want to trace, get its log level.

Since: 1.6.0

pgetConfig :: (Config -> Term s a) -> Term s a

newtype TermMonad m

Constructors

TermMonad

Fields

runTermMonad :: ReaderT Config (Either Text) m

(#) :: Term s (a :--> b) -> Term s a -> Term s b infixl 8

High precedence infixl synonym of papp, to be used like function juxtaposition. e.g.:

>>> f # x # y
f x y

(#$) :: Term s (a :--> b) -> Term s a -> Term s b infixr 0

Low precedence infixr synonym of papp, to be used like $, in combination with #. e.g.:

>>> f # x #$ g # y # z
f x (g y z)