Plutarch.Lift

Description

This module defines functions, associated type families, and newtypes for use with DerivingVia to allow Plutarch to convert to and from PTypes and Haskell types.

Conversion between Plutarch terms and Haskell types

pconstant :: forall p s. PLift p => PLifted p -> Term s p

Create a Plutarch-level constant, from a Haskell value. Example: > pconstant @PInteger 42

plift :: forall p. (HasCallStack, PLift p) => ClosedTerm p -> PLifted p

Like plift' but throws on failure.

plift' :: forall p. PUnsafeLiftDecl p => Config -> ClosedTerm p -> Either LiftError (PLifted p)

Convert a Plutarch term to the associated Haskell value. Fail otherwise. This will fully evaluate the arbitrary closed expression, and convert the resulting value.

data LiftError

Error during script evaluation.

Since: 1.2.1

Constructors

LiftError_EvalError EvalError

LiftError_KnownTypeError BuiltinError

LiftError_FromRepr

LiftError_CompilationError Text

Define your own conversion

class (PUnsafeLiftDecl (PConstanted h), DefaultUni `Includes` PConstantRepr h) => PConstantDecl (h :: Type) where

Class of Haskell types h that can be represented as a Plutus core builtin and converted to a Plutarch type.

The Plutarch type is determined by `PConstanted h`. Its Plutus Core representation is given by `PConstantRepr h`.

This typeclass is closely tied with PLift.

Laws: - pconstantFromRepr . pconstantToRepr ≡ Just - (pconstantToRepr $) . pconstantFromRepr ≡ Just - plift . pfromData . flip ptryFrom fst . pconstant . PlutusTx.toData ≡ id - PlutusTx.fromData . plift . pforgetData . pdata . pconstant ≡ Just

These laws must be upheld for the sake of soundness of the type system.

Associated Types

type PConstantRepr h :: Type

type PConstanted h :: PType

Methods

pconstantToRepr :: h -> PConstantRepr h

pconstantFromRepr :: PConstantRepr h -> Maybe h

type PLift = PUnsafeLiftDecl

Class of Plutarch types p that can be converted to/from a Haskell type.

The Haskell type is determined by `PLifted p`.

This typeclass is closely tied with PConstant.

type PConstant a = (a ~ PLifted (PConstanted a), PConstantDecl a)

Type synonym to simplify deriving of PConstant via DerivePConstantViaNewtype.

A newtype Foo a is considered Constantable if:

The wrapped type a has a PConstant instance.

The lifted type of a has a PUnsafeLiftDecl instance.

There is type equality between a and PLifted (PConstanted a).

These constraints are sufficient to derive a PConstant instance for the newtype.

For deriving PConstant for a wrapped type represented in UPLC as Data, see DerivePConstantViaData.

newtype DerivePConstantDirect (h :: Type) (p :: PType)

Newtype wrapper for deriving PConstant when the wrapped type is directly represented by a builtin UPLC type that is not Data.

Ex: PInteger is directly represented as a builtin integer.

Constructors

DerivePConstantDirect h

newtype DerivePConstantViaNewtype (h :: Type) (p :: PType) (p' :: PType)

Newtype wrapper for deriving PConstant when the wrapped type is represented indirectly by a builtin UPLC type that is not Data.

Ex: PPubKeyHash is a newtype to a PByteString and PByteString is directly represented as a builtin bytestring.

Polymorphic types can be derived as follows:

newtype Foo a = Foo a

newtype PFoo a s = PFoo (Term s a)

instance forall a. PLift a => PUnsafeLiftDecl (PFoo a) where
  type PLifted (PFoo a) = Foo (PLifted a)

deriving via
  ( DerivePConstantViaNewtype
      (Foo a)
      (PFoo (PConstanted a))
      (PConstanted a)
  )
  instance
    PConstant a =>
    PConstantDecl (Foo a)

Constructors

DerivePConstantViaNewtype h
The Haskell newtype we are deriving a PConstant instance for

newtype DerivePConstantViaBuiltin (h :: Type) (p :: PType) (p' :: PType)

Constructors

DerivePConstantViaBuiltin h

Internal use

class (PConstantDecl (PLifted p), PConstanted (PLifted p) ~ p) => PUnsafeLiftDecl (p :: PType)

Laws: - It must be that PConstantRepr (PLifted p) when encoded as a constant in UPLC (via the Constant constructor) is a valid p.

Associated Types

type PLifted p = (r :: Type) | r -> p