Plutarch.Prelude

eDSL types and functions.

data ((a :: PType) :--> (b :: PType)) (s :: S) infixr 0

data PDelayed (a :: PType) (s :: S)

data Term (s :: S) (a :: PType)

type ClosedTerm (a :: PType) = forall (s :: S). Term s a

Closed* terms with no free variables.

plam :: forall c. (PLamN a b s, HasCallStack) => (Term s c -> a) -> Term s (c :--> b)

papp :: Term s (a :--> b) -> Term s a -> Term s b

Lambda Application.

pdelay :: Term s a -> Term s (PDelayed a)

Plutus 'delay', used for laziness.

pforce :: Term s (PDelayed a) -> Term s a

Plutus 'force', used to force evaluation of PDelayed terms.

phoistAcyclic :: HasCallStack => ClosedTerm a -> Term s a

perror :: Term s a

Plutus 'error'.

When using this explicitly, it should be ensured that the containing term is delayed, avoiding premature evaluation.

(#$) :: Term s (a :--> b) -> Term s a -> Term s b infixr 0

Low precedence infixr synonym of papp, to be used like $, in combination with #. e.g.:

>>> f # x #$ g # y # z
f x (g y z)

(#) :: Term s (a :--> b) -> Term s a -> Term s b infixl 8

High precedence infixl synonym of papp, to be used like function juxtaposition. e.g.:

>>> f # x # y
f x y

plet :: Term s a -> (Term s a -> Term s b) -> Term s b

Let bindings.

This is approximately a shorthand for a lambda and application:

plet v f == papp (plam f) v

But sufficiently small terms in WHNF may be inlined for efficiency.

pinl :: Term s a -> (Term s a -> Term s b) -> Term s b

pto :: Term s a -> Term s (PInner a)

Safely coerce from a Term to it's PInner representation.

pfix :: Term s (((a :--> b) :--> (a :--> b)) :--> (a :--> b))

Fixpoint recursion. Used to encode recursive functions.

Example:

iterateN' ::
 Term s (PInteger :--> (a :--> a) :--> a :--> a) ->
 Term s PInteger ->
 Term s (a :--> a) ->
 Term s a
iterateN' self n f x =
  pif (n #== 0) x (self # n - 1 #$ f x)

iterateN :: Term s (PInteger :--> (a :--> a) :--> a :--> a)
iterateN = pfix #$ plam iterateN'

Further examples can be found in examples/Recursion.hs

pthrow :: HasCallStack => Text -> Term s a

type Type = TYPE LiftedRep

The kind of types with lifted values. For example Int :: Type.

data S

type PType = S -> Type

Shorthand for Plutarch types.

class PlutusType (a :: PType)

Associated Types

type PInner a :: PType

type PInner a = DerivedPInner (DPTStrat a) a

class (PInner a ~ DerivedPInner (DPTStrat a) a, PlutusTypeStrat (DPTStrat a), PlutusTypeStratConstraint (DPTStrat a) a, PlutusType a) => DerivePlutusType (a :: PType)

type family DPTStrat a :: Type

data PlutusTypeScott

data PlutusTypeNewtype

data PlutusTypeData

type PCon = PlutusType

Deprecated: Use PlutusType

type PMatch = PlutusType

Deprecated: Use PlutusType

pcon :: PlutusType a => a s -> Term s a

Construct a Plutarch Term via a Haskell datatype

pmatch :: PlutusType a => Term s a -> (a s -> Term s b) -> Term s b

Pattern match over Plutarch Terms via a Haskell datatype

newtype PForall (b :: a -> PType) s

Constructors

PForall (forall (x :: a). Term s (b x))

Integers and integer utilities

data PInteger s

Plutus BuiltinInteger

class PIntegral a where

Minimal complete definition

Nothing

Methods

pdiv :: Term s (a :--> (a :--> a))

default pdiv :: PIntegral (PInner a) => Term s (a :--> (a :--> a))

pmod :: Term s (a :--> (a :--> a))

default pmod :: PIntegral (PInner a) => Term s (a :--> (a :--> a))

pquot :: Term s (a :--> (a :--> a))

default pquot :: PIntegral (PInner a) => Term s (a :--> (a :--> a))

prem :: Term s (a :--> (a :--> a))

default prem :: PIntegral (PInner a) => Term s (a :--> (a :--> a))

Rational numbers and utilities

data PRational s

Note: This type is _not_ the synonym of Rational.

Constructors

PRational (Term s PInteger) (Term s PPositive)

pnumerator :: Term s (PRational :--> PInteger)

pdenominator :: Term s (PRational :--> PPositive)

pround :: Term s (PRational :--> PInteger)

Booleans and boolean functions

data PBool (s :: S)

Plutus BuiltinBool

Constructors

PTrue

PFalse

class PEq t where

Minimal complete definition

Nothing

Methods

(#==) :: Term s t -> Term s t -> Term s PBool infix 4

default (#==) :: (PGeneric t, PlutusType t, All2 PEq (PCode t)) => Term s t -> Term s t -> Term s PBool

class PEq t => PPartialOrd t where

Partial ordering relation.

Laws

#<= must form a partial order. More precisely:

x #<= x = True (#<= is reflexive)

(x #<= y) #&& (y #<= x) = x #== y (#<= is anti-symmetric)

(x #<= y) #&& (y #<= z) = x #<= z (@#<= is transitive)

Furthermore, #< must be an equivalent strict partial order to #<=:

x #<= y = (x #< y) #|| (x #== y)

x #< x = False (#< is irreflexive)

x #< y = pnot (y #< x) (#< is asymmetric)

(x #< y) #&& (y #< z) = x #< z (#< is transitive)

Lastly, if you define #>= or #>, ensure that the following also hold:

x #> y = y #< x

x #>= y = pnot (x #< y)

The default implementations of #>= and #> ensure these laws.

Minimal complete definition

Nothing

Methods

(#<=) :: Term s t -> Term s t -> Term s PBool infix 4

default (#<=) :: POrd (PInner t) => Term s t -> Term s t -> Term s PBool

(#<) :: Term s t -> Term s t -> Term s PBool infix 4

default (#<) :: POrd (PInner t) => Term s t -> Term s t -> Term s PBool

(#>=) :: forall (s :: S). Term s t -> Term s t -> Term s PBool infix 4

@since WIP

(#>) :: forall (s :: S). Term s t -> Term s t -> Term s PBool infix 4

@since WIP

class PPartialOrd t => POrd t where

Total ordering relation.

Laws

pmax and pmin must form a commutative semiring without identity elements, where addition also distributes over multiplication. More precisely:

pmax x y = pmax y x (pmax is commutative)

pmin x y = pmin y x (pmin is commutative)

pmax x (pmax y z) = pmax (pmax x y) z (pmax is associative)

pmin x (pmin y z) = pmin (pmin x y) z (pmin is associative)

pmax x (pmin y z) = pmin (pmax x y) (pmax x z) (pmax distributes over pmin)

pmin x (pmax y z) = pmax (pmin x y) (pmin x z) (pmin distributes over pmax)

Furthermore, the following must hold relative #<:

pmin x y = if (x #< y) then x else y

pmax x y = if (x #< y) then y else x

Laws 7 and 8 are also the defaults, as for most types, this is the best you can do.

Minimal complete definition

Nothing

Methods

pmax :: forall (s :: S). Term s t -> Term s t -> Term s t

@since WIP

pmin :: forall (s :: S). Term s t -> Term s t -> Term s t

@since WIP

pif :: Term s PBool -> Term s a -> Term s a -> Term s a

Lazy if-then-else.

pnot :: Term s (PBool :--> PBool)

Boolean negation for PBool terms.

(#&&) :: Term s PBool -> Term s PBool -> Term s PBool infixr 3

Lazily evaluated boolean and for PBool terms.

(#||) :: Term s PBool -> Term s PBool -> Term s PBool infixr 2

Lazily evaluated boolean or for PBool terms.

Bytestrings and bytestring utilities

Types

data PByteString s

Plutus BuiltinByteString

data PByte (s :: S)

A Plutarch-level representation of bytes.

Note =

This type is intentionally quite restrictive, as it's not really meant to be computed with. Instead, it ensures certain operations' type safety while also allowing more sensible signatures. If you want to do anything with PBytes, we recommend converting them to PIntegers first.

@since WIP

data PLogicOpSemantics (s :: S)

Type designating whether logical operations should use padding or truncation semantics. See CIP-122 for more details on this.

@since WIP

Functions

Construction

ppadding :: forall (s :: S). Term s PLogicOpSemantics

Indicates that padding semantics should be used.

@since WIP

ptruncation :: forall (s :: S). Term s PLogicOpSemantics

Indicates that truncation semantics should be used.

@since WIP

pzeroesBS :: forall (s :: S). Term s (PInteger :--> PByteString)

Construct a PByteString of the specified length (0 if negative) consisting entirely of zero bytes.

@since WIP

ponesBS :: forall (s :: S). Term s (PInteger :--> PByteString)

Construct a PByteString of the specified length (0 if negative) consisting entirely of ones; that is, where every byte is 0xFF.

@since WIP

preplicateBS :: forall (s :: S). Term s (PInteger :--> (PByte :--> PByteString))

Given a desired length and a PByte, construct a PByteString of the specified length (0 if negative) consisting entirely of that PByte.

@since WIP

Byte-oriented

pandBS :: forall (s :: S). Term s (PLogicOpSemantics :--> (PByteString :--> (PByteString :--> PByteString)))

Perform the logical AND of two PByteStrings, as per CIP-122. The PLogicOpSemantics argument specifies what should be done if the lengths of the two PByteString arguments do not match.

@since WIP

porBS :: forall (s :: S). Term s (PLogicOpSemantics :--> (PByteString :--> (PByteString :--> PByteString)))

Perform the logical OR of two PByteStrings, as per CIP-122. The PLogicOpSemantics argument specifies what should be done if the lengths of the two PByteString arguments do not match.

@since WIP

pxorBS :: forall (s :: S). Term s (PLogicOpSemantics :--> (PByteString :--> (PByteString :--> PByteString)))

Perform the logical XOR of two PByteStrings, as per CIP-122. The PLogicOpSemantics argument specifies what should be done if the lengths of the two PByteString arguments do not match.

@since WIP

pcomplementBS :: forall (s :: S). Term s (PByteString :--> PByteString)

Perform the logical complement of a PByteString, as per CIP-122.

@since WIP

pconsBS :: Term s (PByte :--> (PByteString :--> PByteString))

Prepend a PByte to a 'PByteString.

@since WIP

psliceBS :: Term s (PInteger :--> (PInteger :--> (PByteString :--> PByteString)))

Slice a PByteString with given start index and slice length.

>>> (pslice # 2 # 3 phexByteStr "4102afde5b2a") #== phexByteStr "afde5b"

plengthBS :: Term s (PByteString :--> PInteger)

Find the length of a PByteString.

pindexBS :: Term s (PByteString :--> (PInteger :--> PByte))

Given a valid index into a PByteString, returns the PByte at that index. Will crash if given an out-of-bounds index.

@since WIP

Conversion

pbyteToInteger :: Term s (PByte :--> PInteger)

Convert a PByte into its corresponding PInteger.

@since WIP

pintegerToByte :: Term s (PInteger :--> PByte)

Try to convert a PInteger into its corresponding PByte. This operation is checked, and will error if given a negative PInteger, or one too large to fit into a byte.

@since WIP

Other

phexByteStr :: HasCallStack => String -> Term s PByteString

Interpret a hex string as a PByteString.

String and string utilities

data PString s

Plutus BuiltinString values

pencodeUtf8 :: Term s (PString :--> PByteString)

Encode a PString using UTF-8.

pdecodeUtf8 :: Term s (PByteString :--> PString)

Decode a PByteString using UTF-8.

Unit type and utilities

data PUnit s

Constructors

PUnit

Common list typeclass and utilities

class PListLike (list :: PType -> PType) where

Plutarch types that behave like lists.

Minimal complete definition

pelimList, pcons, pnil

Associated Types

type PElemConstraint list (a :: PType) :: Constraint

Methods

pelimList :: PElemConstraint list a => (Term s a -> Term s (list a) -> Term s r) -> Term s r -> Term s (list a) -> Term s r

Canonical eliminator for list-likes.

pcons :: PElemConstraint list a => Term s (a :--> (list a :--> list a))

Cons an element onto an existing list.

pnil :: PElemConstraint list a => Term s (list a)

The empty list

phead :: PElemConstraint list a => Term s (list a :--> a)

Return the first element of a list. Partial, throws an error upon encountering an empty list.

ptail :: PElemConstraint list a => Term s (list a :--> list a)

Take the tail of a list, meaning drop its head. Partial, throws an error upon encountering an empty list.

pnull :: PElemConstraint list a => Term s (list a :--> PBool)

O(1) . Check if a list is empty

type PIsListLike list a = (PListLike list, PElemConstraint list a)

'PIsListLike list a' constraints list be a PListLike with valid element type, a.

plistEquals :: (PIsListLike list a, PEq a) => Term s (list a :--> (list a :--> PBool))

O(min(n, m)) . Check if two lists are equal.

pelem :: (PIsListLike list a, PEq a) => Term s (a :--> (list a :--> PBool))

O(n) . Check if element is in the list

pelemAt :: PIsListLike l a => Term s (PInteger :--> (l a :--> a))

O(n) . Like Haskell level (!!) but on the Plutarch level, not infix and with arguments reversed, errors if the specified index is greater than or equal to the lists length

plength :: PIsListLike list a => Term s (list a :--> PInteger)

O(n) . Count the number of elements in the list

ptryIndex :: PIsListLike list a => Natural -> Term s (list a) -> Term s a

Index a BuiltinList, throwing an error if the index is out of bounds.

pdrop :: PIsListLike list a => Natural -> Term s (list a) -> Term s (list a)

Drop the first n fields of a List.

The term will be statically generated as repeated applications of ptail, which will be more efficient in many circumstances.

psingleton :: PIsListLike list a => Term s (a :--> list a)

O(1) . Create a singleton list from an element

pconcat :: PIsListLike list a => Term s (list a :--> (list a :--> list a))

O(n) . Concatenate two lists

Example: > pconcat # psingleton x # psingleton y == plistLiteral [x, y]

pconcat exhibits identities with empty lists such that > forall x. pconcat # pnil # x == x > forall x. pconcat # x # pnil == x

pzipWith :: (PListLike list, PElemConstraint list a, PElemConstraint list b, PElemConstraint list c) => Term s ((a :--> (b :--> c)) :--> (list a :--> (list b :--> list c)))

O(min(n, m)) . Zip two lists together with a passed function.

If the lists are of differing lengths, cut to the shortest.

pzipWith' :: (PListLike list, PElemConstraint list a, PElemConstraint list b, PElemConstraint list c) => (Term s a -> Term s b -> Term s c) -> Term s (list a :--> (list b :--> list c))

Like pzipWith but with Haskell-level merge function.

pzip :: (PListLike list, PElemConstraint list a, PElemConstraint list b, PElemConstraint list (PPair a b)) => Term s (list a :--> (list b :--> list (PPair a b)))

O(min(n, m)) . Zip two lists together, creating pairs of the elements.

If the lists are of differing lengths, cut to the shortest.

pmap :: (PListLike list, PElemConstraint list a, PElemConstraint list b) => Term s ((a :--> b) :--> (list a :--> list b))

O(n) . Map a function over a list of elements

pfilter :: PIsListLike list a => Term s ((a :--> PBool) :--> (list a :--> list a))

O(n) . Filter elements from a list that don't match the predicate.

pfind :: PIsListLike l a => Term s ((a :--> PBool) :--> (l a :--> PMaybe a))

O(n) . like haskell level find but on plutarch level

precList :: PIsListLike list a => (Term s (list a :--> r) -> Term s a -> Term s (list a) -> Term s r) -> (Term s (list a :--> r) -> Term s r) -> Term s (list a :--> r)

Like pelimList, but with a fixpoint recursion hatch.

pfoldr :: PIsListLike list a => Term s ((a :--> (b :--> b)) :--> (b :--> (list a :--> b)))

O(n) . Fold on a list right-associatively.

pfoldrLazy :: PIsListLike list a => Term s ((a :--> (PDelayed b :--> b)) :--> (b :--> (list a :--> b)))

O(n) . Fold on a list right-associatively, with opportunity for short circuting.

May short circuit when given reducer function is lazy in its second argument.

pfoldl :: PIsListLike list a => Term s ((b :--> (a :--> b)) :--> (b :--> (list a :--> b)))

O(n) . Fold on a list left-associatively.

pall :: PIsListLike list a => Term s ((a :--> PBool) :--> (list a :--> PBool))

O(n) . Check that predicate holds for all elements in a list.

pany :: PIsListLike list a => Term s ((a :--> PBool) :--> (list a :--> PBool))

O(n) . Check that predicate holds for any element in a list.

(#!!) :: PIsListLike l a => Term s (l a) -> Term s PInteger -> Term s a

O(n) . Like Haskell level (!!) but on the plutarch level

Scott encoded list type

data PList (a :: PType) (s :: S)

Constructors

PSCons (Term s a) (Term s (PList a))

PSNil

Scott encoded maybe type and utilities

data PMaybe (a :: S -> Type) (s :: S)

Plutus Maybe type, with Scott-encoded repr

Constructors

PJust (Term s a)

PNothing

Scott encoded either type and utilities

data PEither (a :: PType) (b :: PType) (s :: S)

Scott-encoded Either.

Constructors

PLeft (Term s a)

PRight (Term s b)

Scott encoded pair type and utilities

data PPair (a :: PType) (b :: PType) (s :: S)

Plutus encoding of Pairs.

Note: This is represented differently than BuiltinPair. It is scott-encoded.

Constructors

PPair (Term s a) (Term s b)

Opaque type

newtype POpaque s

An Arbitrary Term with an unknown type

Constructors

POpaque (Term s POpaque)

popaque :: Term s a -> Term s POpaque

Erase the type of a Term

Builtin types and utilities

data PData (s :: S)

pfstBuiltin :: Term s (PBuiltinPair a b :--> a)

psndBuiltin :: Term s (PBuiltinPair a b :--> b)

data PBuiltinPair (a :: PType) (b :: PType) (s :: S)

Plutus BuiltinPair

data PBuiltinList (a :: PType) (s :: S)

Plutus BuiltinList

Constructors

PCons (Term s a) (Term s (PBuiltinList a))

PNil

class PIsData a

Laws: - If PSubtype PData a, then pdataImpl a must be pupcast. - pdataImpl . pupcast . pfromDataImpl ≡ id - pfromDataImpl . punsafeDowncast . pdataImpl ≡ id

pfromData :: PIsData a => Term s (PAsData a) -> Term s a

pdata :: PIsData a => Term s a -> Term s (PAsData a)

data PAsData (a :: PType) (s :: S)

DataRepr and related functions

data PDataRecord (as :: [PLabeledType]) (s :: S)

A "record" of `exists a. PAsData a`. The underlying representation is `PBuiltinList PData`.

data PDataSum defs s

A sum of PDataRecords. The underlying representation is the Constr constructor, where the integer is the index of the variant and the list is the record.

data PLabeledType

Constructors

Symbol := PType

pdcons :: forall label a l s. Term s (PAsData a :--> (PDataRecord l :--> PDataRecord ((label ':= a) ': l)))

Cons a field to a data record.

You can specify the label to associate with the field using type applications-

foo :: Term s (PDataRecord '[ "fooField" ':= PByteString ])
foo = pdcons @"fooField" # pdata (phexByteStr "ab") # pdnil

pdnil :: Term s (PDataRecord '[])

An empty PDataRecord.

pfield :: forall name b p s a as n. (PDataFields p, as ~ PFields p, n ~ PLabelIndex name as, KnownNat n, a ~ PUnLabel (IndexList n as), PFromDataable a b) => Term s (p :--> b)

Get a single field from a Term.

NB*: If you access more than one field from the same value you should use pletFields instead, which will generate the bindings more efficiently.

getField :: HasField x r a => r -> a

Selector function to extract the field from the record.

pletFields :: forall fs a s b ps bs. (PDataFields a, ps ~ PFields a, bs ~ Bindings ps fs, BindFields ps bs) => Term s a -> (HRecOf a fs s -> Term s b) -> Term s b

Bind a HRec of named fields containing all the specified fields.

Tracing

class PShow t

pshow :: PShow a => Term s a -> Term s PString

Return the string representation of a Plutarch value

ptraceInfo :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Trace the given message at the info level before evaluating the given argument.

Since: 1.6.0

ptraceDebug :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Trace the given message at the debug level before evaluating the given argument.

Since: 1.6.0

ptraceInfoShowId :: forall (a :: S -> Type) (s :: S). PShow a => Term s a -> Term s a

Like Haskell's traceShowId but for Plutarch, at the info level.

Since: 1.6.0

ptraceInfoIfFalse :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the info level if the argument is false.

Since: 1.6.0

ptraceInfoIfTrue :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the info level if the argument is true.

Since: 1.6.0

ptraceInfoError :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a

Trace the given message at the info level, then terminate with perror.

Since: 1.6.0

ptraceDebugShowId :: forall (a :: S -> Type) (s :: S). PShow a => Term s a -> Term s a

Like Haskell's traceShowId but for Plutarch, at the debug level.

Since: 1.6.0

ptraceDebugIfFalse :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the debug level if the argument is false.

Since: 1.6.0

ptraceDebugIfTrue :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the debug level if the argument is true.

Since: 1.6.0

ptraceDebugError :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a

Trace the given message at the debug level, then terminate with perror.

Since: 1.6.0

Deprecated

ptrace :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Deprecated: Use ptraceInfo

Backward compatibility synonym for ptraceInfo.

Since: 1.6.0

ptraceShowId :: forall (a :: S -> Type) (s :: S). PShow a => Term s a -> Term s a

Deprecated: Use ptraceInfoShowId

Synonym for ptraceInfoShowId.

Since: 1.6.0

ptraceError :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a

Deprecated: Use ptraceInfoError

Synonym for ptraceInfoError.

Since: 1.6.0

ptraceIfFalse :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Deprecated: Use ptraceInfoIfFalse

Synonym for ptraceInfoIfFalse.

Since: 1.6.0

ptraceIfTrue :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Deprecated: Use ptraceInfoIfTrue

Synonym for ptraceInfoIfTrue.

Since: 1.6.0

Cryptographic hashes and signatures

psha2_256 :: Term s (PByteString :--> PByteString)

Hash a PByteString using SHA-256.

psha3_256 :: Term s (PByteString :--> PByteString)

Hash a PByteString using SHA3-256.

pverifySignature :: Term s (PByteString :--> (PByteString :--> (PByteString :--> PBool)))

Deprecated: use one of the Ed25519, Schnorr- or ECDSA Secp256k1 signature verification functions

Verify the signature against the public key and message.

Conversion between Plutarch terms and Haskell types

pconstant :: forall p s. PLift p => PLifted p -> Term s p

Create a Plutarch-level constant, from a Haskell value. Example: > pconstant @PInteger 42

pconstantData :: forall p h s. (ToData h, PLifted p ~ h, PConstanted h ~ p) => h -> Term s (PAsData p)

Create a Plutarch-level PAsData constant, from a Haskell value. Example: > pconstantData @PInteger 42

plift :: forall p. (HasCallStack, PLift p) => ClosedTerm p -> PLifted p

Like plift' but throws on failure.

type PConstant a = (a ~ PLifted (PConstanted a), PConstantDecl a)

Type synonym to simplify deriving of PConstant via DerivePConstantViaNewtype.

A newtype Foo a is considered Constantable if:

The wrapped type a has a PConstant instance.

The lifted type of a has a PUnsafeLiftDecl instance.

There is type equality between a and PLifted (PConstanted a).

These constraints are sufficient to derive a PConstant instance for the newtype.

For deriving PConstant for a wrapped type represented in UPLC as Data, see DerivePConstantViaData.

type PLift = PUnsafeLiftDecl

Class of Plutarch types p that can be converted to/from a Haskell type.

The Haskell type is determined by `PLifted p`.

This typeclass is closely tied with PConstant.

type PConstantData h = (PConstant h, FromData h, ToData h, PIsData (PConstanted h))

Type synonym to simplify deriving of PConstant via DerivePConstantViaData.

A type Foo a is considered ConstantableData if:

The wrapped type a has a PConstant instance.

The lifted type of a has a PUnsafeLiftDecl instance.

There is type equality between a and PLifted (PConstanted a).

The newtype has FromData and ToData instances

These constraints are sufficient to derive a PConstant instance for the newtype.

For deriving PConstant for a wrapped type represented in UPLC as Data, see DerivePConstantViaData.

Polymorphic types can be derived as follows:

data Bar a = Bar a deriving stock (GHC.Generic)

PlutusTx.makeLift ''Bar
PlutusTx.makeIsDataIndexed ''Bar [('Bar, 0)]

data PBar (a :: PType) (s :: S)
  = PBar (Term s (PDataRecord '["_0" ':= a]))
  deriving stock (GHC.Generic)
  deriving anyclass (SOP.Generic, PIsDataRepr)
  deriving (PlutusType, PIsData, PDataFields) via PIsDataReprInstances (PBar a)

instance
  forall a.
  PLiftData a =>
  PUnsafeLiftDecl (PBar a)
  where
  type PLifted (PBar a) = Bar (PLifted a)

deriving via
  ( DerivePConstantViaData
      (Bar a)
      (PBar (PConstanted a))
  )
  instance
    PConstantData a =>
    PConstantDecl (Bar a)

type PLiftData p = (PLift p, FromData (PLifted p), ToData (PLifted p), PIsData p)

Continuation monad

newtype TermCont :: forall (r :: PType). S -> Type -> Type where

Constructors

TermCont

Fields

:: forall r s a. { runTermCont :: (a -> Term s r) -> Term s r

} -> TermCont @r s a

unTermCont :: TermCont @a s (Term s a) -> Term s a

tcont :: ((a -> Term s r) -> Term s r) -> TermCont @r s a

pletC :: Term s a -> TermCont s (Term s a)

Like plet but works in a TermCont monad

pmatchC :: PlutusType a => Term s a -> TermCont s (a s)

Like pmatch but works in a TermCont monad

pletFieldsC :: forall fs a s b ps bs. (PDataFields a, ps ~ PFields a, bs ~ Bindings ps fs, BindFields ps bs) => Term s a -> TermCont @b s (HRec (BoundTerms ps bs s))

Like pletFields but works in a TermCont monad.

ptraceC :: Term s PString -> TermCont s ()

Like ptrace but works in a TermCont monad.

Example ===

foo :: Term s PUnit
foo = unTermCont $ do
  ptraceC "returning unit!"
  pure $ pconstant ()

pguardC :: Term s PString -> Term s PBool -> TermCont s ()

Trace a message and raise error if cond is false. Otherwise, continue.

Example ===

onlyAllow42 :: Term s (PInteger :--> PUnit)
onlyAllow42 = plam $ i -> unTermCont $ do
  pguardC "expected 42" $ i #== 42
  pure $ pconstant ()

pguardC' :: Term s a -> Term s PBool -> TermCont @a s ()

Stop computation and return given term if cond is false. Otherwise, continue.

Example ===

is42 :: Term s (PInteger :--> PBool)
is42 = plam $ i -> unTermCont $ do
  pguardC' (pconstant False) $ i #== 42
  pure $ pconstant True

ptryFromC :: forall b r a s. PTryFrom a b => Term s a -> TermCont @r s (Term s b, Reduce (PTryFromExcess a b s))

TermCont producing version of ptryFrom.

pupcast :: forall a b s. PSubtype a b => Term s b -> Term s a

ptryFrom :: forall b a s r. PTryFrom a b => Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r

class PSubtype a b => PTryFrom (a :: PType) (b :: PType)

PTryFrom a b represents a subtyping relationship between a and b, and a way to go from a to b. Laws: - (punsafeCoerce . fst) $ tcont (ptryFrom x) ≡ pure x

type family PSubtype (a :: PType) (b :: PType) :: Constraint where ...

Equations

PSubtype a b = (PSubtype' a b ~ 'PSubtypeRelation, PSubtypeHelper a b (PSubtype' a b))

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

A Generic instance must satisfy the following laws:

from . to ≡ id
to . from ≡ id

Minimal complete definition

from, to