Plutarch.Trace

Info level

ptraceInfo :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Trace the given message at the info level before evaluating the given argument.

Since: 1.6.0

ptraceInfoShowId :: forall (a :: S -> Type) (s :: S). PShow a => Term s a -> Term s a

Like Haskell's traceShowId but for Plutarch, at the info level.

Since: 1.6.0

ptraceInfoError :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a

Trace the given message at the info level, then terminate with perror.

Since: 1.6.0

ptraceInfoIfTrue :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the info level if the argument is true.

Since: 1.6.0

ptraceInfoIfFalse :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the info level if the argument is false.

Since: 1.6.0

Debug level

ptraceDebug :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Trace the given message at the debug level before evaluating the given argument.

Since: 1.6.0

ptraceDebugShowId :: forall (a :: S -> Type) (s :: S). PShow a => Term s a -> Term s a

Like Haskell's traceShowId but for Plutarch, at the debug level.

Since: 1.6.0

ptraceDebugIfTrue :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the debug level if the argument is true.

Since: 1.6.0

ptraceDebugIfFalse :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Trace the given message at the debug level if the argument is false.

Since: 1.6.0

ptraceDebugError :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a

Trace the given message at the debug level, then terminate with perror.

Since: 1.6.0

Deprecated

ptrace :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a -> Term s a

Deprecated: Use ptraceInfo

Backward compatibility synonym for ptraceInfo.

Since: 1.6.0

ptraceShowId :: forall (a :: S -> Type) (s :: S). PShow a => Term s a -> Term s a

Deprecated: Use ptraceInfoShowId

Synonym for ptraceInfoShowId.

Since: 1.6.0

ptraceError :: forall (a :: S -> Type) (s :: S). Term s PString -> Term s a

Deprecated: Use ptraceInfoError

Synonym for ptraceInfoError.

Since: 1.6.0

ptraceIfTrue :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Deprecated: Use ptraceInfoIfTrue

Synonym for ptraceInfoIfTrue.

Since: 1.6.0

ptraceIfFalse :: forall (s :: S). Term s PString -> Term s PBool -> Term s PBool

Deprecated: Use ptraceInfoIfFalse

Synonym for ptraceInfoIfFalse.

Since: 1.6.0