Plutarch.TryFrom

Documentation

class PSubtype a b => PTryFrom (a :: PType) (b :: PType) where

PTryFrom a b represents a subtyping relationship between a and b, and a way to go from a to b. Laws: - (punsafeCoerce . fst) $ tcont (ptryFrom x) ≡ pure x

Minimal complete definition

Nothing

Associated Types

type PTryFromExcess a b :: PType

type PTryFromExcess a b = PTryFromExcess a (PInner b)

Methods

ptryFrom' :: forall s r. Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r

default ptryFrom' :: forall s r. (PTryFrom a (PInner b), PTryFromExcess a b ~ PTryFromExcess a (PInner b)) => Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r

ptryFrom :: forall b a s r. PTryFrom a b => Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r

data PSubtypeRelation

Constructors

PSubtypeRelation

PNoSubtypeRelation

type family PSubtype (a :: PType) (b :: PType) :: Constraint where ...

Equations

PSubtype a b = (PSubtype' a b ~ 'PSubtypeRelation, PSubtypeHelper a b (PSubtype' a b))

type family PSubtype' (a :: PType) (b :: PType) :: PSubtypeRelation where ...

Equations

PSubtype' a a = 'PSubtypeRelation

PSubtype' a b = Helper a b (PInner b)

pupcast :: forall a b s. PSubtype a b => Term s b -> Term s a

pupcastF :: forall a b (p :: PType -> PType) s. (PSubtype a b, PCovariant p) => Proxy p -> Term s (p b) -> Term s (p a)

pdowncastF :: forall a b (p :: PType -> PType) s. (PSubtype a b, PContravariant p) => Proxy p -> Term s (p a) -> Term s (p b)